Extract desktop state notification readers
