+++ src/composables/useDesktopStateCore.mts
@@ -1,0 +1,3594 @@
+import { computed, ref } from 'vue'
+import {
+
+  archiveThread,
+  forkThread,
+  getAvailableCollaborationModes,
+  getAccountRateLimits,
+  renameThread,
+  getAvailableModelIds,
+  getCurrentModelConfig,
+  getPendingServerRequests,
+  getSkillsList,
+  getThreadDetail,
+  getOlderThreadMessages,
+  getBackgroundThreadListLimit,
+  interruptThreadTurn,
+  pickCodexRateLimitSnapshot,
+  replyToServerRequest,
+  revertThreadFileChanges,
+  rollbackThread,
+  getThreadGroupsPage,
+  getThreadQueueState,
+  getWorkspaceRootsState,
+  setCodexSpeedMode,
+  setThreadQueueState,
+  setWorkspaceRootsState,
+  getThreadTitleCache,
+  persistThreadTitle,
+  generateThreadTitle,
+  resumeThread,
+
+  startThread,
+  subscribeCodexNotifications,
+  startThreadTurn,
+  type RpcNotification,
+  type SkillInfo,
+  type ThreadQueueState,
+  type WorkspaceRootsState,
+} from '../api/codexGateway'
+import { normalizeFileChangeStatus, toUiFileChanges } from '../api/normalizers/v2'
+import type {
+  CollaborationModeKind,
+  CollaborationModeOption,
+  CommandExecutionData,
+  UiPendingRequestState,
+  ReasoningEffort,
+  SpeedMode,
+  UiFileChange,
+  UiLiveOverlay,
+  UiMessage,
+  UiPlanData,
+  UiPlanStep,
+  UiProjectGroup,
+  UiRateLimitSnapshot,
+  UiServerRequest,
+  UiServerRequestReply,
+  UiThreadTokenUsage,
+  UiTokenUsageBreakdown,
+  UiThread,
+} from '../types/codex'
+import { getPathParent, isProjectlessChatPath, normalizePathForUi, toProjectName } from '../pathUtils.js'
+import {
+  cloneStringKeyedRecord,
+  clamp,
+  createStringKeyedRecord,
+  isThreadUnreadByLastRead,
+  loadProjectDisplayNames,
+  loadProjectOrder,
+  loadReadStateMap,
+  loadSelectedCollaborationModeMap,
+  loadSelectedModelMap,
+  loadSelectedThreadId,
+  loadThreadTerminalOpenMap,
+  loadThreadTokenUsageMap,
+  loadUnreadCutoffIso,
+  NEW_THREAD_COLLABORATION_MODE_CONTEXT,
+  normalizeProviderContextId,
+  normalizeStoredModelId,
+  omitStringKeyedRecordKey,
+  pruneThreadContextStateMap,
+  readSelectedCollaborationMode,
+  readSelectedModel,
+  saveProjectDisplayNames,
+  saveProjectOrder,
+  saveReadStateMap,
+  saveSelectedCollaborationModeMap,
+  saveSelectedModelMap,
+  saveSelectedThreadId,
+  saveThreadTerminalOpenMap,
+  saveThreadTokenUsageMap,
+  saveUnreadCutoffIso,
+  toProviderModelContextId,
+  toThreadContextId,
+  writeSelectedCollaborationModeForContext,
+} from './desktopStateStorage'
+export { isThreadUnreadByLastRead } from './desktopStateStorage'
+import {
+  WORKED_MESSAGE_TYPE,
+  BACKGROUND_THREAD_PAGINATION_DELAY_MS,
+  EVENT_SYNC_DEBOUNCE_MS,
+  GLOBAL_SERVER_REQUEST_SCOPE,
+  MODEL_FALLBACK_ID,
+  RATE_LIMIT_REFRESH_DEBOUNCE_MS,
+  REASONING_EFFORT_OPTIONS,
+  RECENT_THREAD_MESSAGE_LOAD_REUSE_MS,
+  TURN_START_FOLLOW_UP_SYNC_DELAY_MS,
+  addWorkspaceRootPlaceholderGroups,
+  areGroupArraysEqual,
+  areMessageArraysEqual,
+  areStringArraysEqual,
+  areThreadArraysEqual,
+  areThreadFieldsEqual,
+  areTurnActivitiesEqual,
+  areTurnSummariesEqual,
+  buildWorkspaceRootsProjectOrderState,
+  collectWorkspaceRootPathsForProjectRemoval,
+  collectDuplicateProjectLeafNames,
+  disambiguateProjectGroupsByCwd,
+  filterGroupsByWorkspaceRoots,
+  findAdjacentThreadId,
+  flattenThreads,
+  formatTurnDuration,
+  getRemoteProjectDisplayName,
+  getWorkspaceProjectOrderNames,
+  getWorkspaceProjectOrderPaths,
+  insertTurnSummaryMessage,
+  isProjectlessGroup,
+  isUnsupportedChatGptModelError,
+  matchesWorkspaceRootProject,
+  mergeIncomingWithLocalInProgressThreads,
+  mergeMessages,
+  mergeProjectOrder,
+  normalizeMessageText,
+  mergeThreadGroups,
+  omitKey,
+  omitKeys,
+  orderGroupsByProjectOrder,
+  orderGroupsByWorkspaceProjectOrder,
+  parseIsoTimestamp,
+  pruneThreadStateMap,
+  removePersistedLiveMessages,
+  removeRedundantLiveAgentMessages,
+  removeThreadFromGroups,
+  reorderStringArray,
+  toForkedThreadTitle,
+  toOptimisticThreadTitle,
+  toProjectNameFromWorkspaceRoot,
+  upsertMessage,
+  type TurnActivityState,
+  type TurnCompletedInfo,
+  type TurnErrorState,
+  type TurnStartedInfo,
+  type TurnSummaryState,
+} from './desktopStateThreadHelpers'
+import {
+  asRecord,
+  buildPlanMessageText,
+  extractThreadIdFromNotification,
+  isAgentContentEvent,
+  liveReasoningMessageId,
+  normalizePlanStepStatus,
+  normalizeRateLimitSnapshotsPayload,
+  normalizeServerRequest,
+  readAgentMessageCompleted,
+  readAgentMessageDelta,
+  readAgentMessageStartedId,
+  readCommandOutputDelta,
+  readCompletedImageView,
+  readNotificationErrorState,
+  readNumber,
+  readReasoningCompletedId,
+  readReasoningDelta,
+  readReasoningSectionBreakMessageId,
+  readReasoningStartedItemId,
+  readString,
+  readThreadTokenUsageUpdate,
+  readToolRequestUserInputQuestionIds,
+  readTurnActivity,
+  readTurnCompletedInfo,
+  readTurnErrorMessage,
+  readTurnStartedInfo,
+  sanitizeDisplayText,
+} from './desktopStateNotificationReaders'
+export {
+  buildWorkspaceRootsProjectOrderState,
+  collectWorkspaceRootPathsForProjectRemoval,
+  filterGroupsByWorkspaceRoots,
+  findAdjacentThreadId,
+  removeThreadFromGroups,
+} from './desktopStateThreadHelpers'
+
+export function useDesktopState() {
+  const projectGroups = ref<UiProjectGroup[]>([])
+  const sourceGroups = ref<UiProjectGroup[]>([])
+  const selectedThreadId = ref(loadSelectedThreadId())
+  const persistedMessagesByThreadId = ref<Record<string, UiMessage[]>>({})
+  const livePlanMessagesByThreadId = ref<Record<string, UiMessage[]>>({})
+  const liveAgentMessagesByThreadId = ref<Record<string, UiMessage[]>>({})
+  const liveReasoningTextByThreadId = ref<Record<string, string>>({})
+  const liveCommandsByThreadId = ref<Record<string, UiMessage[]>>({})
+  const liveFileChangeMessagesByThreadId = ref<Record<string, UiMessage[]>>({})
+  const inProgressById = ref<Record<string, boolean>>({})
+  type FileAttachment = { label: string; path: string; fsPath: string }
+  type QueuedMessage = {
+    id: string
+    text: string
+    imageUrls: string[]
+    skills: Array<{ name: string; path: string }>
+    fileAttachments: FileAttachment[]
+    collaborationMode: CollaborationModeKind
+  }
+  type PendingTurnRequest = {
+    text: string
+    imageUrls: string[]
+    skills: Array<{ name: string; path: string }>
+    fileAttachments: FileAttachment[]
+    effort: ReasoningEffort | ''
+    collaborationMode: CollaborationModeKind
+    fallbackRetried: boolean
+  }
+  const queuedMessagesByThreadId = ref<Record<string, QueuedMessage[]>>({})
+  const queueProcessingByThreadId = ref<Record<string, boolean>>({})
+  let hasLoadedPersistedQueueState = false
+  const eventUnreadByThreadId = ref<Record<string, boolean>>({})
+  const availableModelIds = ref<string[]>([])
+  const availableCollaborationModes = ref<CollaborationModeOption[]>([
+    { value: 'default', label: 'Default' },
+    { value: 'plan', label: 'Plan' },
+  ])
+  const selectedCollaborationModeByContext = ref<Record<string, CollaborationModeKind>>(
+    loadSelectedCollaborationModeMap(),
+  )
+  const selectedModelIdByContext = ref<Record<string, string>>(loadSelectedModelMap())
+  const selectedCollaborationMode = ref<CollaborationModeKind>(
+    readSelectedCollaborationMode(selectedCollaborationModeByContext.value, selectedThreadId.value),
+  )
+  const selectedModelId = ref(readSelectedModel(selectedModelIdByContext.value, selectedThreadId.value))
+  const selectedReasoningEffort = ref<ReasoningEffort | ''>('medium')
+  const selectedSpeedMode = ref<SpeedMode>('standard')
+  const activeProviderId = ref('')
+  const readStateByThreadId = ref<Record<string, string>>(loadReadStateMap())
+  const unreadCutoffIso = ref(loadUnreadCutoffIso())
+  const projectOrder = ref<string[]>(loadProjectOrder())
+  const projectDisplayNameById = ref<Record<string, string>>(loadProjectDisplayNames())
+  const loadedVersionByThreadId = ref<Record<string, string>>({})
+  const loadedMessagesByThreadId = ref<Record<string, boolean>>({})
+  const hasMoreOlderMessagesByThreadId = ref<Record<string, boolean>>({})
+  const loadingOlderMessagesByThreadId = ref<Record<string, boolean>>({})
+  const resumedThreadById = ref<Record<string, boolean>>({})
+  const turnIndexByTurnIdByThreadId = ref<Record<string, Record<string, number>>>({})
+  const turnSummaryByThreadId = ref<Record<string, TurnSummaryState>>({})
+  const turnActivityByThreadId = ref<Record<string, TurnActivityState>>({})
+  const turnErrorByThreadId = ref<Record<string, TurnErrorState>>({})
+  const activeTurnIdByThreadId = ref<Record<string, string>>({})
+  const interruptBlockedUntilPersistedByThreadId = ref<Record<string, boolean>>({})
+  const threadListedByServerById = ref<Record<string, boolean>>({})
+  const persistedUserMessageByThreadId = ref<Record<string, boolean>>({})
+  const pendingServerRequestsByThreadId = ref<Record<string, UiServerRequest[]>>({})
+  const pendingTurnRequestByThreadId = ref<Record<string, PendingTurnRequest>>({})
+  const codexRateLimit = ref<UiRateLimitSnapshot | null>(null)
+  const threadTokenUsageByThreadId = ref<Record<string, UiThreadTokenUsage>>(loadThreadTokenUsageMap())
+  const terminalOpenByThreadId = ref<Record<string, boolean>>(loadThreadTerminalOpenMap())
+
+  const threadTitleById = ref<Record<string, string>>({})
+
+  const installedSkills = ref<SkillInfo[]>([])
+  const accountRateLimitSnapshots = ref<UiRateLimitSnapshot[]>([])
+
+  const isLoadingThreads = ref(false)
+  const isLoadingMessages = ref(false)
+  const isThreadListFullyLoaded = ref(false)
+  const isSendingMessage = ref(false)
+  const isInterruptingTurn = ref(false)
+  const isUpdatingSpeedMode = ref(false)
+  const isRollingBack = ref(false)
+
+  const error = ref('')
+  const isPolling = ref(false)
+  const hasLoadedThreads = ref(false)
+
+  function extractLocalImagePathFromUrl(value: string): string {
+    try {
+      const parsed = new URL(value, 'http://localhost')
+      if (parsed.pathname !== '/codex-local-image') return ''
+      return parsed.searchParams.get('path')?.trim() ?? ''
+    } catch {
+      return ''
+    }
+  }
+
+  function shouldReuseAttachedImageFromPrompt(promptText: string): boolean {
+    const normalized = promptText.trim().toLowerCase()
+    if (!normalized) return false
+    return /\b(attached image|attached screenshot|save the attached|copy (the )?screenshot|save screenshot)\b/i.test(normalized)
+  }
+
+  function findLatestUserLocalImageUrl(threadId: string): string {
+    const persisted = persistedMessagesByThreadId.value[threadId] ?? []
+    for (let index = persisted.length - 1; index >= 0; index -= 1) {
+      const message = persisted[index]
+      if (message.role !== 'user' || !Array.isArray(message.images) || message.images.length === 0) continue
+      for (let imageIndex = message.images.length - 1; imageIndex >= 0; imageIndex -= 1) {
+        const imageUrl = message.images[imageIndex]?.trim() ?? ''
+        if (!imageUrl) continue
+        if (extractLocalImagePathFromUrl(imageUrl)) return imageUrl
+      }
+    }
+    return ''
+  }
+  let stopNotificationStream: (() => void) | null = null
+  let eventSyncTimer: number | null = null
+  let rateLimitRefreshTimer: number | null = null
+  const delayedTurnSyncTimerByThreadId = new Map<string, number>()
+  let loadThreadsPromise: Promise<void> | null = null
+  const loadMessagePromiseByThreadId = new Map<string, Promise<void>>()
+  let refreshSkillsPromise: Promise<void> | null = null
+  let rateLimitRefreshPromise: Promise<void> | null = null
+  let pendingThreadsRefresh = false
+  const pendingThreadMessageRefresh = new Set<string>()
+  const lastMessageLoadAtByThreadId = new Map<string, number>()
+  let threadListNextCursor: string | null = null
+  let threadListBackgroundTimer: number | null = null
+  let isLoadingRemainingThreadPages = false
+  let hasLoadedAllThreadPages = false
+  let loadedThreadListGroups: UiProjectGroup[] = []
+  let loadedThreadListRootsState: WorkspaceRootsState | null = null
+  let hasHydratedWorkspaceRootsState = false
+  let activeReasoningItemId = ''
+  let shouldAutoScrollOnNextAgentEvent = false
+  const pendingTurnStartsById = new Map<string, TurnStartedInfo>()
+  const fallbackRetryInFlightThreadIds = new Set<string>()
+
+
+  const allThreads = computed(() => flattenThreads(projectGroups.value))
+  const selectedThread = computed(() =>
+    allThreads.value.find((thread) => thread.id === selectedThreadId.value) ?? null,
+  )
+  const selectedThreadTerminalOpen = computed(() => {
+    const threadId = selectedThreadId.value
+    return Boolean(threadId && terminalOpenByThreadId.value[threadId] === true)
+  })
+  const isSelectedThreadInterruptPending = computed(() => {
+    const threadId = selectedThreadId.value
+    if (!threadId) return false
+    return interruptBlockedUntilPersistedByThreadId.value[threadId] === true
+  })
+  const selectedThreadServerRequests = computed<UiServerRequest[]>(() => {
+    const rows: UiServerRequest[] = []
+    const selected = selectedThreadId.value
+    if (selected && Array.isArray(pendingServerRequestsByThreadId.value[selected])) {
+      rows.push(...pendingServerRequestsByThreadId.value[selected])
+    }
+    if (Array.isArray(pendingServerRequestsByThreadId.value[GLOBAL_SERVER_REQUEST_SCOPE])) {
+      rows.push(...pendingServerRequestsByThreadId.value[GLOBAL_SERVER_REQUEST_SCOPE])
+    }
+    return rows.sort((first, second) => first.receivedAtIso.localeCompare(second.receivedAtIso))
+  })
+  const selectedLiveOverlay = computed<UiLiveOverlay | null>(() => {
+    const threadId = selectedThreadId.value
+    if (!threadId) return null
+
+    const isInProgress = inProgressById.value[threadId] === true
+    const activity = isInProgress ? turnActivityByThreadId.value[threadId] : undefined
+    const reasoningText = isInProgress
+      ? (liveReasoningTextByThreadId.value[threadId] ?? '').trim()
+      : ''
+    const errorText = (turnErrorByThreadId.value[threadId]?.message ?? '').trim()
+
+    if (!activity && !reasoningText && !errorText) return null
+    return {
+      activityLabel: activity?.label || 'Thinking',
+      activityDetails: activity?.details ?? [],
+      reasoningText,
+      errorText,
+    }
+  })
+  const codexQuota = computed<UiRateLimitSnapshot | null>(() => codexRateLimit.value)
+  const selectedThreadTokenUsage = computed<UiThreadTokenUsage | null>(() => {
+    const threadId = selectedThreadId.value
+    if (!threadId) return null
+    return threadTokenUsageByThreadId.value[threadId] ?? null
+  })
+  const messages = computed<UiMessage[]>(() => {
+    const threadId = selectedThreadId.value
+    if (!threadId) return []
+
+    const persisted = persistedMessagesByThreadId.value[threadId] ?? []
+    const livePlan = livePlanMessagesByThreadId.value[threadId] ?? []
+    const liveAgent = liveAgentMessagesByThreadId.value[threadId] ?? []
+    const liveCommands = liveCommandsByThreadId.value[threadId] ?? []
+    const liveFileChanges = liveFileChangeMessagesByThreadId.value[threadId] ?? []
+    const combined = [...persisted, ...livePlan, ...liveCommands, ...liveFileChanges, ...liveAgent]
+
+    const summary = turnSummaryByThreadId.value[threadId]
+    if (!summary) return combined
+    return insertTurnSummaryMessage(combined, summary)
+  })
+  const hasMoreOlderMessages = computed(() => {
+    const threadId = selectedThreadId.value
+    return threadId ? hasMoreOlderMessagesByThreadId.value[threadId] === true : false
+  })
+  const isLoadingOlderMessages = computed(() => {
+    const threadId = selectedThreadId.value
+    return threadId ? loadingOlderMessagesByThreadId.value[threadId] === true : false
+  })
+
+  function getFirstPersistedTurnId(threadId: string): string {
+    const persisted = persistedMessagesByThreadId.value[threadId] ?? []
+    for (const message of persisted) {
+      const turnId = message.turnId?.trim() ?? ''
+      if (turnId) return turnId
+    }
+    return ''
+  }
+
+  function readModelIdForThread(threadId: string): string {
+    const contextId = toThreadContextId(threadId)
+    if (contextId === NEW_THREAD_COLLABORATION_MODE_CONTEXT) {
+      const providerContextId = toProviderModelContextId(activeProviderId.value)
+      const providerModelId = providerContextId
+        ? normalizeStoredModelId(selectedModelIdByContext.value[providerContextId])
+        : ''
+      if (providerModelId) return providerModelId
+    }
+    return readSelectedModel(selectedModelIdByContext.value, threadId).trim()
+  }
+
+  function ensureAvailableModelIds(...modelIds: string[]): void {
+    const nextModelIds = [...availableModelIds.value]
+    for (const modelId of modelIds) {
+      const normalizedModelId = modelId.trim()
+      if (normalizedModelId && !nextModelIds.includes(normalizedModelId)) {
+        nextModelIds.push(normalizedModelId)
+      }
+    }
+    if (!areStringArraysEqual(availableModelIds.value, nextModelIds)) {
+      availableModelIds.value = nextModelIds
+    }
+  }
+
+  function setSelectedThreadId(nextThreadId: string): void {
+    if (selectedThreadId.value === nextThreadId) return
+    selectedThreadId.value = nextThreadId
+    saveSelectedThreadId(nextThreadId)
+    selectedModelId.value = readModelIdForThread(nextThreadId)
+    ensureAvailableModelIds(selectedModelId.value)
+    selectedCollaborationMode.value = readSelectedCollaborationMode(
+      selectedCollaborationModeByContext.value,
+      nextThreadId,
+    )
+    activeReasoningItemId = ''
+    shouldAutoScrollOnNextAgentEvent = false
+  }
+
+  function setSelectedModelIdForThread(threadId: string, modelId: string): void {
+    const normalizedModelId = modelId.trim()
+    const contextId = toThreadContextId(threadId)
+    if (normalizedModelId) {
+      const nextModelMap = cloneStringKeyedRecord(selectedModelIdByContext.value)
+      nextModelMap[contextId] = normalizedModelId
+      selectedModelIdByContext.value = nextModelMap
+    } else {
+      selectedModelIdByContext.value = omitStringKeyedRecordKey(selectedModelIdByContext.value, contextId)
+    }
+    if (threadId.trim() === selectedThreadId.value) {
+      selectedModelId.value = readModelIdForThread(selectedThreadId.value)
+      ensureAvailableModelIds(selectedModelId.value)
+    } else {
+      ensureAvailableModelIds(normalizedModelId)
+    }
+    if (contextId === NEW_THREAD_COLLABORATION_MODE_CONTEXT) {
+      const providerContextId = toProviderModelContextId(activeProviderId.value)
+      if (providerContextId) {
+        if (normalizedModelId) {
+          const nextModelMap = cloneStringKeyedRecord(selectedModelIdByContext.value)
+          nextModelMap[providerContextId] = normalizedModelId
+          selectedModelIdByContext.value = nextModelMap
+        } else {
+          selectedModelIdByContext.value = omitStringKeyedRecordKey(selectedModelIdByContext.value, providerContextId)
+        }
+      }
+    }
+    saveSelectedModelMap(selectedModelIdByContext.value)
+  }
+
+  function setSelectedModelId(modelId: string): void {
+    setSelectedModelIdForThread(selectedThreadId.value, modelId)
+  }
+
+  function setThreadModelId(threadId: string, modelId: string): void {
+    const normalizedThreadId = threadId.trim()
+    if (!normalizedThreadId) return
+
+    const normalizedModelId = modelId.trim()
+    if (normalizedModelId) {
+      const nextModelMap = cloneStringKeyedRecord(selectedModelIdByContext.value)
+      nextModelMap[normalizedThreadId] = normalizedModelId
+      selectedModelIdByContext.value = nextModelMap
+    } else {
+      selectedModelIdByContext.value = omitStringKeyedRecordKey(selectedModelIdByContext.value, normalizedThreadId)
+    }
+    ensureAvailableModelIds(normalizedModelId)
+    if (selectedThreadId.value === normalizedThreadId) {
+      selectedModelId.value = readModelIdForThread(selectedThreadId.value)
+    }
+    saveSelectedModelMap(selectedModelIdByContext.value)
+  }
+
+  function setThreadTokenUsage(threadId: string, usage: UiThreadTokenUsage | null): void {
+    const normalizedThreadId = threadId.trim()
+    if (!normalizedThreadId) return
+
+    if (!usage) {
+      if (!(normalizedThreadId in threadTokenUsageByThreadId.value)) return
+      threadTokenUsageByThreadId.value = omitKey(threadTokenUsageByThreadId.value, normalizedThreadId)
+      saveThreadTokenUsageMap(threadTokenUsageByThreadId.value)
+      return
+    }
+
+    const current = threadTokenUsageByThreadId.value[normalizedThreadId]
+    if (current && JSON.stringify(current) === JSON.stringify(usage)) return
+
+    threadTokenUsageByThreadId.value = {
+      ...threadTokenUsageByThreadId.value,
+      [normalizedThreadId]: usage,
+    }
+    saveThreadTokenUsageMap(threadTokenUsageByThreadId.value)
+  }
+
+  function setSelectedCollaborationMode(mode: CollaborationModeKind): void {
+    const nextMode: CollaborationModeKind = mode === 'plan' ? 'plan' : 'default'
+    const contextId = toThreadContextId(selectedThreadId.value)
+    const currentMode = readSelectedCollaborationMode(selectedCollaborationModeByContext.value, selectedThreadId.value)
+    if (currentMode === nextMode && selectedCollaborationMode.value === nextMode) return
+    selectedCollaborationMode.value = nextMode
+    selectedCollaborationModeByContext.value = writeSelectedCollaborationModeForContext(
+      selectedCollaborationModeByContext.value,
+      contextId,
+      nextMode,
+    )
+    saveSelectedCollaborationModeMap(selectedCollaborationModeByContext.value)
+  }
+
+  function setSelectedCollaborationModeForThread(threadId: string, mode: CollaborationModeKind): void {
+    const nextMode = mode === 'plan' ? 'plan' : 'default'
+    selectedCollaborationModeByContext.value = writeSelectedCollaborationModeForContext(
+      selectedCollaborationModeByContext.value,
+      threadId,
+      nextMode,
+    )
+    if (threadId.trim() === selectedThreadId.value) {
+      selectedCollaborationMode.value = nextMode
+    }
+    saveSelectedCollaborationModeMap(selectedCollaborationModeByContext.value)
+  }
+
+  function setCodexRateLimit(nextSnapshot: UiRateLimitSnapshot | null): void {
+    codexRateLimit.value = nextSnapshot
+  }
+
+  async function applyFallbackModelSelection(threadId: string = selectedThreadId.value): Promise<void> {
+    if (threadId.trim()) {
+      setThreadModelId(threadId, MODEL_FALLBACK_ID)
+    } else {
+      setSelectedModelId(MODEL_FALLBACK_ID)
+    }
+    ensureAvailableModelIds(MODEL_FALLBACK_ID)
+  }
+
+  function setPendingTurnRequest(threadId: string, request: PendingTurnRequest): void {
+    pendingTurnRequestByThreadId.value = {
+      ...pendingTurnRequestByThreadId.value,
+      [threadId]: request,
+    }
+  }
+
+  function clearPendingTurnRequest(threadId: string): void {
+    if (!pendingTurnRequestByThreadId.value[threadId]) return
+    pendingTurnRequestByThreadId.value = omitKey(pendingTurnRequestByThreadId.value, threadId)
+  }
+
+
+
+  async function retryPendingTurnWithFallback(threadId: string): Promise<void> {
+    if (fallbackRetryInFlightThreadIds.has(threadId)) return
+    const pending = pendingTurnRequestByThreadId.value[threadId]
+    if (!pending || pending.fallbackRetried) return
+
+    fallbackRetryInFlightThreadIds.add(threadId)
+    setPendingTurnRequest(threadId, {
+      ...pending,
+      fallbackRetried: true,
+    })
+
+    try {
+      await applyFallbackModelSelection(threadId)
+      // Remove the failed user turn before replaying on fallback model to avoid duplicated user messages.
+      try {
+        const rolledBackMessages = await rollbackThread(threadId, 1)
+        setPersistedMessagesForThread(threadId, rolledBackMessages)
+        clearLivePlansForThread(threadId)
+        setLiveAgentMessagesForThread(threadId, [])
+        clearLiveReasoningForThread(threadId)
+        if (liveCommandsByThreadId.value[threadId]) {
+          liveCommandsByThreadId.value = omitKey(liveCommandsByThreadId.value, threadId)
+        }
+      } catch {
+        // If rollback fails, continue with retry rather than dropping the turn.
+      }
+      setTurnErrorForThread(threadId, null)
+      error.value = ''
+      setTurnSummaryForThread(threadId, null)
+      setTurnActivityForThread(threadId, {
+        label: 'Thinking',
+        details: buildPendingTurnDetails(MODEL_FALLBACK_ID, pending.effort, pending.collaborationMode),
+      })
+      setThreadInProgress(threadId, true)
+
+      if (resumedThreadById.value[threadId] !== true) {
+        await resumeThread(threadId)
+      }
+
+      await startThreadTurn(
+        threadId,
+        pending.text,
+        pending.imageUrls,
+        MODEL_FALLBACK_ID,
+        pending.effort || undefined,
+        pending.skills.length > 0 ? pending.skills : undefined,
+        pending.fileAttachments,
+        pending.collaborationMode,
+      )
+
+      resumedThreadById.value = {
+        ...resumedThreadById.value,
+        [threadId]: true,
+      }
+
+      scheduleRateLimitRefresh()
+      pendingThreadMessageRefresh.add(threadId)
+      await syncFromNotifications()
+    } catch (unknownError) {
+      const errorMessage = unknownError instanceof Error ? unknownError.message : 'Unknown application error'
+      setTurnErrorForThread(threadId, errorMessage)
+      error.value = errorMessage
+      setThreadInProgress(threadId, false)
+      setTurnActivityForThread(threadId, null)
+    } finally {
+      fallbackRetryInFlightThreadIds.delete(threadId)
+    }
+  }
+
+  function setSelectedReasoningEffort(effort: ReasoningEffort | ''): void {
+    if (effort && !REASONING_EFFORT_OPTIONS.includes(effort)) {
+      return
+    }
+    selectedReasoningEffort.value = effort
+  }
+
+  async function updateSelectedSpeedMode(mode: SpeedMode): Promise<void> {
+    const nextMode: SpeedMode = mode === 'fast' ? 'fast' : 'standard'
+    if (isUpdatingSpeedMode.value || selectedSpeedMode.value === nextMode) {
+      return
+    }
+
+    const previousMode = selectedSpeedMode.value
+    selectedSpeedMode.value = nextMode
+    isUpdatingSpeedMode.value = true
+    error.value = ''
+
+    try {
+      await setCodexSpeedMode(nextMode)
+    } catch (unknownError) {
+      selectedSpeedMode.value = previousMode
+      error.value = unknownError instanceof Error ? unknownError.message : 'Failed to update Fast mode'
+    } finally {
+      isUpdatingSpeedMode.value = false
+    }
+  }
+
+  async function refreshCollaborationModes(): Promise<void> {
+    try {
+      const modes = await getAvailableCollaborationModes()
+      availableCollaborationModes.value = modes
+      if (!modes.some((mode) => mode.value === selectedCollaborationMode.value)) {
+        setSelectedCollaborationMode('default')
+      }
+    } catch {
+      // Keep the last known collaboration mode choices on transient failures.
+    }
+  }
+
+  function buildPendingTurnDetails(
+    modelId: string,
+    effort: ReasoningEffort | '',
+    collaborationMode: CollaborationModeKind = selectedCollaborationMode.value,
+  ): string[] {
+    const modelLabel = modelId.trim() || 'default'
+    const effortLabel = effort || 'default'
+    const modeLabel = collaborationMode === 'plan' ? 'Plan' : 'Default'
+    const speedLabel = selectedSpeedMode.value === 'fast' ? 'Fast' : 'Standard'
+    return [`Mode: ${modeLabel}`, `Model: ${modelLabel}`, `Thinking: ${effortLabel}`, `Speed: ${speedLabel}`]
+  }
+
+  async function refreshModelPreferences(options?: { providerChanged?: boolean; includeProviderModels?: boolean }): Promise<void> {
+    try {
+      const [modelIds, currentConfig] = await Promise.all([
+        getAvailableModelIds({ includeProviderModels: options?.includeProviderModels !== false }),
+        getCurrentModelConfig(),
+      ])
+
+      const normalizedSelectedModelId = readModelIdForThread(selectedThreadId.value)
+      const normalizedConfiguredModelId = currentConfig.model.trim()
+      const normalizedProviderId = normalizeProviderContextId(currentConfig.providerId)
+      activeProviderId.value = normalizedProviderId
+      const providerModelContextId = toProviderModelContextId(normalizedProviderId)
+      const providerScopedModelId = providerModelContextId
+        ? normalizeStoredModelId(selectedModelIdByContext.value[providerModelContextId])
+        : ''
+      const nextModelIds = [...modelIds]
+      if (!options?.providerChanged) {
+        for (const modelId of [normalizedSelectedModelId, normalizedConfiguredModelId]) {
+          if (modelId && !nextModelIds.includes(modelId)) {
+            nextModelIds.push(modelId)
+          }
+        }
+      }
+      availableModelIds.value = nextModelIds
+
+      const currentModelInNewList = normalizedSelectedModelId && modelIds.includes(normalizedSelectedModelId)
+      if (!normalizedSelectedModelId || !currentModelInNewList || options?.providerChanged) {
+        if (options?.providerChanged && nextModelIds.length > 0) {
+          if (providerScopedModelId && nextModelIds.includes(providerScopedModelId)) {
+            setSelectedModelId(providerScopedModelId)
+          } else if (normalizedConfiguredModelId && nextModelIds.includes(normalizedConfiguredModelId)) {
+            setSelectedModelId(normalizedConfiguredModelId)
+          } else {
+            setSelectedModelId(nextModelIds[0])
+          }
+        } else if (normalizedConfiguredModelId && nextModelIds.includes(normalizedConfiguredModelId)) {
+          setSelectedModelId(currentConfig.model)
+        } else if (nextModelIds.length > 0) {
+          setSelectedModelId(nextModelIds[0])
+        } else {
+          setSelectedModelId('')
+        }
+      }
+      if (providerModelContextId && selectedModelId.value.trim().length > 0) {
+        const nextModelMap = cloneStringKeyedRecord(selectedModelIdByContext.value)
+        nextModelMap[providerModelContextId] = selectedModelId.value.trim()
+        selectedModelIdByContext.value = nextModelMap
+        saveSelectedModelMap(selectedModelIdByContext.value)
+      }
+
+      if (
+        currentConfig.reasoningEffort &&
+        REASONING_EFFORT_OPTIONS.includes(currentConfig.reasoningEffort)
+      ) {
+        selectedReasoningEffort.value = currentConfig.reasoningEffort
+      }
+      selectedSpeedMode.value = currentConfig.speedMode
+    } catch {
+      // Keep chat UI usable even if model metadata is temporarily unavailable.
+    }
+  }
+
+  async function refreshRateLimits(): Promise<void> {
+    if (rateLimitRefreshPromise) {
+      await rateLimitRefreshPromise
+      return
+    }
+
+    rateLimitRefreshPromise = (async () => {
+      try {
+        const snapshot = await getAccountRateLimits()
+        setCodexRateLimit(snapshot)
+        accountRateLimitSnapshots.value = snapshot ? [snapshot] : []
+      } catch {
+        // Keep the last known rate-limit state if the endpoint is temporarily unavailable.
+      } finally {
+        rateLimitRefreshPromise = null
+      }
+    })()
+
+    await rateLimitRefreshPromise
+  }
+
+  function scheduleRateLimitRefresh(): void {
+    if (typeof window === 'undefined') {
+      void refreshRateLimits()
+      return
+    }
+
+    if (rateLimitRefreshTimer !== null) {
+      window.clearTimeout(rateLimitRefreshTimer)
+    }
+
+    rateLimitRefreshTimer = window.setTimeout(() => {
+      rateLimitRefreshTimer = null
+      void refreshRateLimits()
+    }, RATE_LIMIT_REFRESH_DEBOUNCE_MS)
+  }
+
+  function clearDelayedTurnSync(threadId: string): void {
+    if (!threadId || typeof window === 'undefined') return
+    const timerId = delayedTurnSyncTimerByThreadId.get(threadId)
+    if (timerId === undefined) return
+    window.clearTimeout(timerId)
+    delayedTurnSyncTimerByThreadId.delete(threadId)
+  }
+
+  function scheduleDelayedTurnSync(threadId: string): void {
+    if (!threadId || typeof window === 'undefined') return
+    clearDelayedTurnSync(threadId)
+    const timerId = window.setTimeout(() => {
+      delayedTurnSyncTimerByThreadId.delete(threadId)
+      pendingThreadMessageRefresh.add(threadId)
+      void syncFromNotifications()
+    }, TURN_START_FOLLOW_UP_SYNC_DELAY_MS)
+    delayedTurnSyncTimerByThreadId.set(threadId, timerId)
+  }
+
+  function applyCachedTitlesToGroups(groups: UiProjectGroup[]): UiProjectGroup[] {
+    const titles = threadTitleById.value
+    if (Object.keys(titles).length === 0) return groups
+    return groups.map((group) => ({
+      projectName: group.projectName,
+      threads: group.threads.map((thread) => {
+        const cached = titles[thread.id]
+        return cached ? { ...thread, title: cached } : thread
+      }),
+    }))
+  }
+
+  function getThreadPendingRequests(threadId: string): UiServerRequest[] {
+    if (!threadId) return []
+    return Array.isArray(pendingServerRequestsByThreadId.value[threadId])
+      ? pendingServerRequestsByThreadId.value[threadId]
+      : []
+  }
+
+  function isApprovalRequestMethod(method: string): boolean {
+    return (
+      method === 'item/commandExecution/requestApproval' ||
+      method === 'item/fileChange/requestApproval' ||
+      method === 'item/permissions/requestApproval' ||
+      method === 'execCommandApproval' ||
+      method === 'applyPatchApproval'
+    )
+  }
+
+  function readPendingRequestState(requests: UiServerRequest[]): UiPendingRequestState | null {
+    if (requests.some((request) => isApprovalRequestMethod(request.method))) {
+      return 'approval'
+    }
+    return requests.length > 0 ? 'response' : null
+  }
+
+  function applyThreadFlags(): void {
+    const withTitles = applyCachedTitlesToGroups(sourceGroups.value)
+    const flaggedGroups: UiProjectGroup[] = withTitles.map((group) => ({
+      projectName: group.projectName,
+      threads: group.threads.map((thread) => {
+        const inProgress = inProgressById.value[thread.id] === true
+        const pendingRequestState = readPendingRequestState(getThreadPendingRequests(thread.id))
+        const isSelected = selectedThreadId.value === thread.id
+        const unreadByEvent = eventUnreadByThreadId.value[thread.id] === true
+        const unreadByTime = isThreadUnreadByLastRead(
+          thread.updatedAtIso,
+          readStateByThreadId.value[thread.id],
+          unreadCutoffIso.value,
+        )
+        const unread = !isSelected && !inProgress && (unreadByEvent || unreadByTime)
+
+        return {
+          ...thread,
+          inProgress,
+          unread,
+          pendingRequestState,
+        }
+      }),
+    }))
+    projectGroups.value = mergeThreadGroups(projectGroups.value, flaggedGroups)
+  }
+
+  function insertOptimisticThread(threadId: string, cwd: string, firstMessageText: string): void {
+    const nowIso = new Date().toISOString()
+    const normalizedCwd = normalizePathForUi(cwd)
+    const projectName = toProjectName(normalizedCwd)
+    const nextThread: UiThread = {
+      id: threadId,
+      title: toOptimisticThreadTitle(firstMessageText),
+      projectName,
+      cwd: normalizedCwd,
+      hasWorktree: normalizedCwd.includes('/.codex/worktrees/') || normalizedCwd.includes('/.git/worktrees/'),
+      createdAtIso: nowIso,
+      updatedAtIso: nowIso,
+      preview: firstMessageText,
+      unread: false,
+      inProgress: false,
+    }
+
+    const existingGroupIndex = sourceGroups.value.findIndex((group) => group.projectName === projectName)
+    if (existingGroupIndex >= 0) {
+      const existingGroup = sourceGroups.value[existingGroupIndex]
+      const remainingThreads = existingGroup.threads.filter((thread) => thread.id !== threadId)
+      const nextGroup: UiProjectGroup = {
+        projectName,
+        threads: [nextThread, ...remainingThreads],
+      }
+      const nextGroups = [...sourceGroups.value]
+      nextGroups.splice(existingGroupIndex, 1, nextGroup)
+      sourceGroups.value = nextGroups
+    } else {
+      sourceGroups.value = [{ projectName, threads: [nextThread] }, ...sourceGroups.value]
+    }
+
+    const nextProjectOrder = mergeProjectOrder(projectOrder.value, sourceGroups.value)
+    if (!areStringArraysEqual(projectOrder.value, nextProjectOrder)) {
+      projectOrder.value = nextProjectOrder
+      saveProjectOrder(projectOrder.value)
+    }
+    applyThreadFlags()
+  }
+
+  function pruneThreadScopedState(flatThreads: UiThread[]): void {
+    const activeThreadIds = new Set(flatThreads.map((thread) => thread.id))
+    const currentThreadId = selectedThreadId.value.trim()
+    if (currentThreadId) {
+      activeThreadIds.add(currentThreadId)
+    }
+    const nextSelectedModelMap = pruneThreadContextStateMap(selectedModelIdByContext.value, activeThreadIds)
+    if (nextSelectedModelMap !== selectedModelIdByContext.value) {
+      selectedModelIdByContext.value = nextSelectedModelMap
+      selectedModelId.value = readModelIdForThread(selectedThreadId.value)
+      ensureAvailableModelIds(selectedModelId.value)
+      saveSelectedModelMap(nextSelectedModelMap)
+    }
+    const nextSelectedCollaborationModeMap = pruneThreadContextStateMap(
+      selectedCollaborationModeByContext.value,
+      activeThreadIds,
+    )
+    if (nextSelectedCollaborationModeMap !== selectedCollaborationModeByContext.value) {
+      selectedCollaborationModeByContext.value = nextSelectedCollaborationModeMap
+      selectedCollaborationMode.value = readSelectedCollaborationMode(
+        nextSelectedCollaborationModeMap,
+        selectedThreadId.value,
+      )
+      saveSelectedCollaborationModeMap(nextSelectedCollaborationModeMap)
+    }
+    const nextReadState = pruneThreadStateMap(readStateByThreadId.value, activeThreadIds)
+    if (nextReadState !== readStateByThreadId.value) {
+      readStateByThreadId.value = nextReadState
+      saveReadStateMap(nextReadState)
+    }
+    loadedMessagesByThreadId.value = pruneThreadStateMap(loadedMessagesByThreadId.value, activeThreadIds)
+    loadedVersionByThreadId.value = pruneThreadStateMap(loadedVersionByThreadId.value, activeThreadIds)
+    resumedThreadById.value = pruneThreadStateMap(resumedThreadById.value, activeThreadIds)
+    turnIndexByTurnIdByThreadId.value = pruneThreadStateMap(turnIndexByTurnIdByThreadId.value, activeThreadIds)
+    persistedMessagesByThreadId.value = pruneThreadStateMap(persistedMessagesByThreadId.value, activeThreadIds)
+    liveAgentMessagesByThreadId.value = pruneThreadStateMap(liveAgentMessagesByThreadId.value, activeThreadIds)
+    liveReasoningTextByThreadId.value = pruneThreadStateMap(liveReasoningTextByThreadId.value, activeThreadIds)
+    liveCommandsByThreadId.value = pruneThreadStateMap(liveCommandsByThreadId.value, activeThreadIds)
+    liveFileChangeMessagesByThreadId.value = pruneThreadStateMap(liveFileChangeMessagesByThreadId.value, activeThreadIds)
+    turnSummaryByThreadId.value = pruneThreadStateMap(turnSummaryByThreadId.value, activeThreadIds)
+    turnActivityByThreadId.value = pruneThreadStateMap(turnActivityByThreadId.value, activeThreadIds)
+    turnErrorByThreadId.value = pruneThreadStateMap(turnErrorByThreadId.value, activeThreadIds)
+    activeTurnIdByThreadId.value = pruneThreadStateMap(activeTurnIdByThreadId.value, activeThreadIds)
+    interruptBlockedUntilPersistedByThreadId.value = pruneThreadStateMap(
+      interruptBlockedUntilPersistedByThreadId.value,
+      activeThreadIds,
+    )
+    threadListedByServerById.value = pruneThreadStateMap(threadListedByServerById.value, activeThreadIds)
+    persistedUserMessageByThreadId.value = pruneThreadStateMap(persistedUserMessageByThreadId.value, activeThreadIds)
+    const nextQueuedMessages = pruneThreadStateMap(queuedMessagesByThreadId.value, activeThreadIds)
+    if (nextQueuedMessages !== queuedMessagesByThreadId.value) {
+      queuedMessagesByThreadId.value = nextQueuedMessages
+      persistQueueState()
+    }
+    threadTokenUsageByThreadId.value = pruneThreadStateMap(threadTokenUsageByThreadId.value, activeThreadIds)
+    eventUnreadByThreadId.value = pruneThreadStateMap(eventUnreadByThreadId.value, activeThreadIds)
+    inProgressById.value = pruneThreadStateMap(inProgressById.value, activeThreadIds)
+    const nextPending: Record<string, UiServerRequest[]> = {}
+    for (const [threadId, requests] of Object.entries(pendingServerRequestsByThreadId.value)) {
+      if (threadId === GLOBAL_SERVER_REQUEST_SCOPE || activeThreadIds.has(threadId)) {
+        nextPending[threadId] = requests
+      }
+    }
+    pendingServerRequestsByThreadId.value = nextPending
+  }
+
+  function markThreadAsRead(threadId: string): void {
+    const thread = flattenThreads(sourceGroups.value).find((row) => row.id === threadId)
+    if (!thread) return
+
+    readStateByThreadId.value = {
+      ...readStateByThreadId.value,
+      [threadId]: thread.updatedAtIso,
+    }
+    saveReadStateMap(readStateByThreadId.value)
+    if (eventUnreadByThreadId.value[threadId]) {
+      eventUnreadByThreadId.value = omitKey(eventUnreadByThreadId.value, threadId)
+    }
+    applyThreadFlags()
+  }
+
+  function setTurnSummaryForThread(threadId: string, summary: TurnSummaryState | null): void {
+    if (!threadId) return
+
+    const previous = turnSummaryByThreadId.value[threadId]
+    if (summary) {
+      if (areTurnSummariesEqual(previous, summary)) return
+      turnSummaryByThreadId.value = {
+        ...turnSummaryByThreadId.value,
+        [threadId]: summary,
+      }
+    } else {
+      if (previous) {
+        turnSummaryByThreadId.value = omitKey(turnSummaryByThreadId.value, threadId)
+      }
+    }
+  }
+
+  function setThreadInProgress(threadId: string, nextInProgress: boolean): void {
+    if (!threadId) return
+    const currentValue = inProgressById.value[threadId] === true
+    if (currentValue === nextInProgress) return
+    if (nextInProgress) {
+      inProgressById.value = {
+        ...inProgressById.value,
+        [threadId]: true,
+      }
+    } else {
+      inProgressById.value = omitKey(inProgressById.value, threadId)
+      clearCompletedTurnLiveState(threadId)
+      clearInterruptPersistenceGate(threadId)
+    }
+    applyThreadFlags()
+    if (!nextInProgress && !hasActiveInProgressThreads() && threadListNextCursor) {
+      scheduleRemainingThreadPages()
+    }
+  }
+
+  function clearInterruptPersistenceGate(threadId: string): void {
+    if (!threadId) return
+    if (interruptBlockedUntilPersistedByThreadId.value[threadId]) {
+      interruptBlockedUntilPersistedByThreadId.value = omitKey(interruptBlockedUntilPersistedByThreadId.value, threadId)
+    }
+    if (threadListedByServerById.value[threadId]) {
+      threadListedByServerById.value = omitKey(threadListedByServerById.value, threadId)
+    }
+    if (persistedUserMessageByThreadId.value[threadId]) {
+      persistedUserMessageByThreadId.value = omitKey(persistedUserMessageByThreadId.value, threadId)
+    }
+  }
+
+  function blockInterruptUntilThreadIsPersisted(threadId: string): void {
+    if (!threadId) return
+    interruptBlockedUntilPersistedByThreadId.value = {
+      ...interruptBlockedUntilPersistedByThreadId.value,
+      [threadId]: true,
+    }
+    if (threadListedByServerById.value[threadId]) {
+      threadListedByServerById.value = omitKey(threadListedByServerById.value, threadId)
+    }
+    if (persistedUserMessageByThreadId.value[threadId]) {
+      persistedUserMessageByThreadId.value = omitKey(persistedUserMessageByThreadId.value, threadId)
+    }
+  }
+
+  function maybeUnblockInterruptForPersistedThread(threadId: string): void {
+    if (!threadId) return
+    if (interruptBlockedUntilPersistedByThreadId.value[threadId] !== true) return
+    if (threadListedByServerById.value[threadId] !== true) return
+    if (persistedUserMessageByThreadId.value[threadId] !== true) return
+    clearInterruptPersistenceGate(threadId)
+  }
+
+  function maybeUnblockInterruptForActiveTurn(threadId: string, turnId: string): void {
+    if (!threadId || !turnId) return
+    if (interruptBlockedUntilPersistedByThreadId.value[threadId] !== true) return
+    clearInterruptPersistenceGate(threadId)
+  }
+
+  function markServerListedThreads(serverThreadIds: Set<string>): void {
+    const pendingThreadIds = Object.keys(interruptBlockedUntilPersistedByThreadId.value)
+    if (pendingThreadIds.length === 0) return
+
+    let nextListedState = threadListedByServerById.value
+    let changed = false
+    for (const threadId of pendingThreadIds) {
+      if (!serverThreadIds.has(threadId) || nextListedState[threadId] === true) continue
+      nextListedState = {
+        ...nextListedState,
+        [threadId]: true,
+      }
+      changed = true
+    }
+
+    if (!changed) return
+    threadListedByServerById.value = nextListedState
+    for (const threadId of pendingThreadIds) {
+      maybeUnblockInterruptForPersistedThread(threadId)
+    }
+  }
+
+  function markThreadMessagesPersisted(threadId: string, messages: UiMessage[]): void {
+    if (!threadId) return
+    if (interruptBlockedUntilPersistedByThreadId.value[threadId] !== true) return
+    if (!messages.some((message) => message.role === 'user')) return
+    if (persistedUserMessageByThreadId.value[threadId] !== true) {
+      persistedUserMessageByThreadId.value = {
+        ...persistedUserMessageByThreadId.value,
+        [threadId]: true,
+      }
+    }
+    maybeUnblockInterruptForPersistedThread(threadId)
+  }
+
+  function markThreadUnreadByEvent(threadId: string): void {
+    if (!threadId) return
+    if (threadId === selectedThreadId.value) return
+    if (eventUnreadByThreadId.value[threadId] === true) return
+    eventUnreadByThreadId.value = {
+      ...eventUnreadByThreadId.value,
+      [threadId]: true,
+    }
+    applyThreadFlags()
+  }
+
+  function setTurnActivityForThread(threadId: string, activity: TurnActivityState | null): void {
+    if (!threadId) return
+
+    const previous = turnActivityByThreadId.value[threadId]
+    if (!activity) {
+      if (previous) {
+        turnActivityByThreadId.value = omitKey(turnActivityByThreadId.value, threadId)
+      }
+      return
+    }
+
+    const normalizedLabel = sanitizeDisplayText(activity.label) || 'Thinking'
+    const incomingDetails = activity.details
+      .map((line) => sanitizeDisplayText(line))
+      .filter((line) => line.length > 0 && line !== normalizedLabel)
+    const mergedDetails = Array.from(new Set([...(previous?.details ?? []), ...incomingDetails])).slice(-3)
+    const nextActivity: TurnActivityState = {
+      label: normalizedLabel,
+      details: mergedDetails,
+    }
+
+    if (areTurnActivitiesEqual(previous, nextActivity)) return
+    turnActivityByThreadId.value = {
+      ...turnActivityByThreadId.value,
+      [threadId]: nextActivity,
+    }
+  }
+
+  function setTurnErrorForThread(
+    threadId: string,
+    message: string | null,
+    options: { transient?: boolean } = {},
+  ): void {
+    if (!threadId) return
+
+    const previous = turnErrorByThreadId.value[threadId]
+    const normalizedMessage = message ? normalizeMessageText(message) : ''
+    if (!normalizedMessage) {
+      if (previous) {
+        turnErrorByThreadId.value = omitKey(turnErrorByThreadId.value, threadId)
+      }
+      return
+    }
+
+    const transient = options.transient === true
+    if (previous?.message === normalizedMessage && previous.transient === transient) return
+
+    turnErrorByThreadId.value = {
+      ...turnErrorByThreadId.value,
+      [threadId]: { message: normalizedMessage, transient },
+    }
+  }
+
+  function clearTransientTurnErrorForThread(threadId: string): void {
+    if (!threadId) return
+    if (!turnErrorByThreadId.value[threadId]?.transient) return
+    setTurnErrorForThread(threadId, null)
+  }
+
+  function clearAllTransientTurnErrors(): void {
+    const transientThreadIds = Object.entries(turnErrorByThreadId.value)
+      .filter(([, state]) => state?.transient)
+      .map(([threadId]) => threadId)
+    if (transientThreadIds.length === 0) return
+
+    let nextState = turnErrorByThreadId.value
+    for (const threadId of transientThreadIds) {
+      nextState = omitKey(nextState, threadId)
+    }
+    turnErrorByThreadId.value = nextState
+  }
+
+  function currentThreadVersion(threadId: string): string {
+    const thread = flattenThreads(sourceGroups.value).find((row) => row.id === threadId)
+    return thread?.updatedAtIso ?? ''
+  }
+
+  function setThreadTerminalOpen(threadId: string, isOpen: boolean): void {
+    if (!threadId) return
+    const next = { ...terminalOpenByThreadId.value }
+    if (isOpen) {
+      next[threadId] = true
+    } else {
+      delete next[threadId]
+    }
+    terminalOpenByThreadId.value = next
+    saveThreadTerminalOpenMap(next)
+  }
+
+  function toggleSelectedThreadTerminal(): void {
+    const threadId = selectedThreadId.value
+    if (!threadId) return
+    setThreadTerminalOpen(threadId, !selectedThreadTerminalOpen.value)
+  }
+
+  function setPersistedMessagesForThread(threadId: string, nextMessages: UiMessage[]): void {
+    const previous = persistedMessagesByThreadId.value[threadId] ?? []
+    if (areMessageArraysEqual(previous, nextMessages)) return
+    persistedMessagesByThreadId.value = {
+      ...persistedMessagesByThreadId.value,
+      [threadId]: nextMessages,
+    }
+  }
+
+  function setLiveAgentMessagesForThread(threadId: string, nextMessages: UiMessage[]): void {
+    const previous = liveAgentMessagesByThreadId.value[threadId] ?? []
+    if (areMessageArraysEqual(previous, nextMessages)) return
+    liveAgentMessagesByThreadId.value = {
+      ...liveAgentMessagesByThreadId.value,
+      [threadId]: nextMessages,
+    }
+  }
+
+  function clearLiveAgentMessagesForThread(threadId: string): void {
+    if (!threadId) return
+    if (!(threadId in liveAgentMessagesByThreadId.value)) return
+    liveAgentMessagesByThreadId.value = omitKey(liveAgentMessagesByThreadId.value, threadId)
+  }
+
+  function setLiveFileChangeMessagesForThread(threadId: string, nextMessages: UiMessage[]): void {
+    const previous = liveFileChangeMessagesByThreadId.value[threadId] ?? []
+    if (areMessageArraysEqual(previous, nextMessages)) return
+    liveFileChangeMessagesByThreadId.value = {
+      ...liveFileChangeMessagesByThreadId.value,
+      [threadId]: nextMessages,
+    }
+  }
+
+  function setLivePlanMessagesForThread(threadId: string, nextMessages: UiMessage[]): void {
+    const previous = livePlanMessagesByThreadId.value[threadId] ?? []
+    if (areMessageArraysEqual(previous, nextMessages)) return
+    livePlanMessagesByThreadId.value = {
+      ...livePlanMessagesByThreadId.value,
+      [threadId]: nextMessages,
+    }
+  }
+
+  function upsertLivePlanMessage(threadId: string, nextMessage: UiMessage): void {
+    const previous = livePlanMessagesByThreadId.value[threadId] ?? []
+    const next = upsertMessage(previous, nextMessage)
+    setLivePlanMessagesForThread(threadId, next)
+  }
+
+  function upsertLiveAgentMessage(threadId: string, nextMessage: UiMessage): void {
+    const previous = liveAgentMessagesByThreadId.value[threadId] ?? []
+    const next = upsertMessage(previous, nextMessage)
+    setLiveAgentMessagesForThread(threadId, next)
+  }
+
+  function upsertLiveFileChangeMessage(threadId: string, nextMessage: UiMessage): void {
+    const previous = liveFileChangeMessagesByThreadId.value[threadId] ?? []
+    const next = upsertMessage(previous, nextMessage)
+    setLiveFileChangeMessagesForThread(threadId, next)
+  }
+
+  function setLiveReasoningText(threadId: string, text: string): void {
+    if (!threadId) return
+    const normalized = text.trim()
+    const previous = liveReasoningTextByThreadId.value[threadId] ?? ''
+    if (normalized.length === 0) {
+      if (!previous) return
+      liveReasoningTextByThreadId.value = omitKey(liveReasoningTextByThreadId.value, threadId)
+      return
+    }
+    if (previous === normalized) return
+    liveReasoningTextByThreadId.value = {
+      ...liveReasoningTextByThreadId.value,
+      [threadId]: normalized,
+    }
+  }
+
+  function appendLiveReasoningText(threadId: string, delta: string): void {
+    if (!threadId) return
+    const previous = liveReasoningTextByThreadId.value[threadId] ?? ''
+    setLiveReasoningText(threadId, `${previous}${delta}`)
+  }
+
+  function clearLiveReasoningForThread(threadId: string): void {
+    if (!threadId) return
+    if (!(threadId in liveReasoningTextByThreadId.value)) return
+    liveReasoningTextByThreadId.value = omitKey(liveReasoningTextByThreadId.value, threadId)
+  }
+
+  function clearLivePlansForThread(threadId: string): void {
+    if (!threadId) return
+    if (!(threadId in livePlanMessagesByThreadId.value)) return
+    livePlanMessagesByThreadId.value = omitKey(livePlanMessagesByThreadId.value, threadId)
+  }
+
+  function clearLiveFileChangesForThread(threadId: string): void {
+    if (!threadId) return
+    if (!(threadId in liveFileChangeMessagesByThreadId.value)) return
+    liveFileChangeMessagesByThreadId.value = omitKey(liveFileChangeMessagesByThreadId.value, threadId)
+  }
+
+  function clearCompletedTurnLiveState(threadId: string): void {
+    if (!threadId) return
+    clearLivePlansForThread(threadId)
+    clearLiveReasoningForThread(threadId)
+    setTurnActivityForThread(threadId, null)
+    if (threadId === selectedThreadId.value) {
+      activeReasoningItemId = ''
+    }
+    if (liveCommandsByThreadId.value[threadId]) {
+      liveCommandsByThreadId.value = omitKey(liveCommandsByThreadId.value, threadId)
+    }
+    if (activeTurnIdByThreadId.value[threadId]) {
+      activeTurnIdByThreadId.value = omitKey(activeTurnIdByThreadId.value, threadId)
+    }
+    clearPendingTurnRequest(threadId)
+  }
+
+  function readPlanUpdate(notification: RpcNotification): { threadId: string; message: UiMessage } | null {
+    if (notification.method !== 'turn/plan/updated') return null
+    const params = asRecord(notification.params)
+    const threadId = extractThreadIdFromNotification(notification)
+    const turnId = readString(params?.turnId) || readString(params?.turn_id)
+    const rawSteps = Array.isArray(params?.plan) ? params?.plan : []
+    const steps: UiPlanStep[] = rawSteps
+      .map((row) => asRecord(row))
+      .map((row) => ({
+        step: readString(row?.step),
+        status: normalizePlanStepStatus(row?.status),
+      }))
+      .filter((row) => row.step.length > 0)
+
+    if (!threadId || !turnId) return null
+
+    const explanation = readString(params?.explanation).trim()
+    const plan: UiPlanData = {
+      explanation: explanation || undefined,
+      steps,
+      isStreaming: true,
+    }
+
+    return {
+      threadId,
+      message: {
+        id: `${turnId}:plan`,
+        role: 'assistant',
+        text: buildPlanMessageText(plan),
+        messageType: 'plan.live',
+        plan,
+      },
+    }
+  }
+
+  function readPlanDelta(notification: RpcNotification): { threadId: string; message: UiMessage } | null {
+    if (notification.method !== 'item/plan/delta') return null
+    const params = asRecord(notification.params)
+    const threadId = extractThreadIdFromNotification(notification)
+    const turnId = readString(params?.turnId) || readString(params?.turn_id)
+    const delta = readString(params?.delta)
+    if (!threadId || !turnId || !delta) return null
+
+    const messageId = `${turnId}:plan`
+    const existing = (livePlanMessagesByThreadId.value[threadId] ?? []).find((message) => message.id === messageId)
+    const nextText = `${existing?.text ?? ''}${delta}`
+    const nextPlan: UiPlanData | undefined = existing?.plan
+      ? { ...existing.plan, isStreaming: true }
+      : undefined
+
+    return {
+      threadId,
+      message: {
+        id: messageId,
+        role: 'assistant',
+        text: nextText,
+        messageType: 'plan.live',
+        plan: nextPlan,
+      },
+    }
+  }
+
+  function upsertPendingServerRequest(request: UiServerRequest): void {
+    const threadId = request.threadId || GLOBAL_SERVER_REQUEST_SCOPE
+    const current = pendingServerRequestsByThreadId.value[threadId] ?? []
+    const index = current.findIndex((row) => row.id === request.id)
+    const nextRows = [...current]
+    if (index >= 0) {
+      nextRows.splice(index, 1, request)
+    } else {
+      nextRows.push(request)
+    }
+
+    pendingServerRequestsByThreadId.value = {
+      ...pendingServerRequestsByThreadId.value,
+      [threadId]: nextRows.sort((first, second) => first.receivedAtIso.localeCompare(second.receivedAtIso)),
+    }
+    applyThreadFlags()
+  }
+
+  function removePendingServerRequestById(requestId: number): void {
+    const next: Record<string, UiServerRequest[]> = {}
+    for (const [threadId, requests] of Object.entries(pendingServerRequestsByThreadId.value)) {
+      const filtered = requests.filter((request) => request.id !== requestId)
+      if (filtered.length > 0) {
+        next[threadId] = filtered
+      }
+    }
+    pendingServerRequestsByThreadId.value = next
+    applyThreadFlags()
+  }
+
+  function replacePendingServerRequests(requests: UiServerRequest[]): void {
+    const next: Record<string, UiServerRequest[]> = {}
+    for (const request of requests) {
+      const threadId = request.threadId || GLOBAL_SERVER_REQUEST_SCOPE
+      const current = next[threadId] ?? []
+      current.push(request)
+      next[threadId] = current
+    }
+
+    for (const rows of Object.values(next)) {
+      rows.sort((first, second) => first.receivedAtIso.localeCompare(second.receivedAtIso))
+    }
+
+    pendingServerRequestsByThreadId.value = next
+  }
+
+  function handleServerRequestNotification(notification: RpcNotification): boolean {
+    if (notification.method === 'server/request') {
+      const request = normalizeServerRequest(notification.params)
+      if (!request) return true
+      upsertPendingServerRequest(request)
+      return true
+    }
+
+    if (notification.method === 'server/request/resolved') {
+      const row = asRecord(notification.params)
+      const id = row?.id
+      if (typeof id === 'number' && Number.isInteger(id)) {
+        removePendingServerRequestById(id)
+      }
+      return true
+    }
+
+    return false
+  }
+
+  function inferNextTurnIndex(threadId: string): number {
+    const persisted = persistedMessagesByThreadId.value[threadId] ?? []
+    let maxTurnIndex = -1
+    for (const message of persisted) {
+      if (typeof message.turnIndex === 'number' && Number.isFinite(message.turnIndex)) {
+        maxTurnIndex = Math.max(maxTurnIndex, message.turnIndex)
+      }
+    }
+    return maxTurnIndex + 1
+  }
+
+  function setTurnIndexForThread(threadId: string, turnId: string, turnIndex: number): void {
+    if (!threadId || !turnId || !Number.isInteger(turnIndex) || turnIndex < 0) return
+    const previous = turnIndexByTurnIdByThreadId.value[threadId] ?? {}
+    if (previous[turnId] === turnIndex) return
+    turnIndexByTurnIdByThreadId.value = {
+      ...turnIndexByTurnIdByThreadId.value,
+      [threadId]: {
+        ...previous,
+        [turnId]: turnIndex,
+      },
+    }
+  }
+
+  function replaceTurnIndexLookupForThread(threadId: string, nextLookup: Record<string, number>): void {
+    const previous = turnIndexByTurnIdByThreadId.value[threadId] ?? {}
+    const previousEntries = Object.entries(previous)
+    const nextEntries = Object.entries(nextLookup)
+    if (
+      previousEntries.length === nextEntries.length
+      && previousEntries.every(([turnId, turnIndex]) => nextLookup[turnId] === turnIndex)
+    ) {
+      return
+    }
+
+    turnIndexByTurnIdByThreadId.value = {
+      ...turnIndexByTurnIdByThreadId.value,
+      [threadId]: { ...nextLookup },
+    }
+  }
+
+  function rebindLiveFileChangeTurnIndices(threadId: string): void {
+    const current = liveFileChangeMessagesByThreadId.value[threadId]
+    if (!current || current.length === 0) return
+
+    const turnIndexByTurnId = turnIndexByTurnIdByThreadId.value[threadId] ?? {}
+    let changed = false
+    const next = current.map((message) => {
+      if (typeof message.turnIndex === 'number' || !message.turnId) {
+        return message
+      }
+      const turnIndex = turnIndexByTurnId[message.turnId]
+      if (typeof turnIndex !== 'number') return message
+      changed = true
+      return { ...message, turnIndex }
+    })
+
+    if (!changed) return
+    liveFileChangeMessagesByThreadId.value = {
+      ...liveFileChangeMessagesByThreadId.value,
+      [threadId]: next,
+    }
+  }
+
+  function readCommandExecutionStarted(notification: RpcNotification): UiMessage | null {
+    if (notification.method !== 'item/started') return null
+    const params = asRecord(notification.params)
+    const item = asRecord(params?.item)
+    if (!item || item.type !== 'commandExecution') return null
+    const id = readString(item.id)
+    const command = readString(item.command)
+    if (!id) return null
+    const cwd = typeof item.cwd === 'string' ? item.cwd : null
+    const threadId = extractThreadIdFromNotification(notification)
+    const turnId = readString(params?.turnId) || readString(params?.turn_id)
+    const turnIndex = threadId && turnId
+      ? turnIndexByTurnIdByThreadId.value[threadId]?.[turnId]
+      : undefined
+    return {
+      id,
+      role: 'system',
+      text: command,
+      messageType: 'commandExecution',
+      commandExecution: { command, cwd, status: 'inProgress', aggregatedOutput: '', exitCode: null },
+      turnId: turnId || undefined,
+      turnIndex: typeof turnIndex === 'number' ? turnIndex : undefined,
+    }
+  }
+
+  function readCommandExecutionCompleted(notification: RpcNotification): UiMessage | null {
+    if (notification.method !== 'item/completed') return null
+    const params = asRecord(notification.params)
+    const item = asRecord(params?.item)
+    if (!item || item.type !== 'commandExecution') return null
+    const id = readString(item.id)
+    const command = readString(item.command)
+    if (!id) return null
+    const cwd = typeof item.cwd === 'string' ? item.cwd : null
+    const statusRaw = readString(item.status)
+    const status: CommandExecutionData['status'] =
+      statusRaw === 'failed' ? 'failed' : statusRaw === 'declined' ? 'declined' : statusRaw === 'interrupted' ? 'interrupted' : 'completed'
+    const aggregatedOutput = typeof item.aggregatedOutput === 'string' ? item.aggregatedOutput : ''
+    const exitCode = typeof item.exitCode === 'number' ? item.exitCode : null
+    const threadId = extractThreadIdFromNotification(notification)
+    const turnId = readString(params?.turnId) || readString(params?.turn_id)
+    const turnIndex = threadId && turnId
+      ? turnIndexByTurnIdByThreadId.value[threadId]?.[turnId]
+      : undefined
+    return {
+      id,
+      role: 'system',
+      text: command,
+      messageType: 'commandExecution',
+      commandExecution: { command, cwd, status, aggregatedOutput, exitCode },
+      turnId: turnId || undefined,
+      turnIndex: typeof turnIndex === 'number' ? turnIndex : undefined,
+    }
+  }
+
+  function readCompletedFileChange(notification: RpcNotification): UiMessage | null {
+    if (notification.method !== 'item/completed') return null
+    const params = asRecord(notification.params)
+    const item = asRecord(params?.item)
+    if (!item || item.type !== 'fileChange') return null
+    const id = readString(item.id)
+    if (!id) return null
+    const threadId = readString(params?.threadId)
+    const turnId = readString(params?.turnId)
+    const turnIndex = threadId && turnId
+      ? turnIndexByTurnIdByThreadId.value[threadId]?.[turnId]
+      : undefined
+
+    const fileChanges = toUiFileChanges(item.changes)
+    const fileChangeStatus = normalizeFileChangeStatus(item.status)
+    if (fileChanges.length === 0 || fileChangeStatus !== 'completed') return null
+
+    return {
+      id,
+      role: 'system',
+      text: '',
+      messageType: 'fileChange',
+      fileChangeStatus,
+      fileChanges,
+      turnId: turnId || undefined,
+      turnIndex: typeof turnIndex === 'number' ? turnIndex : undefined,
+    }
+  }
+
+  function upsertLiveCommand(threadId: string, msg: UiMessage): void {
+    const previous = liveCommandsByThreadId.value[threadId] ?? []
+    const next = upsertMessage(previous, msg)
+    if (next === previous) return
+    liveCommandsByThreadId.value = { ...liveCommandsByThreadId.value, [threadId]: next }
+  }
+
+  function removeLiveCommandsPersistedIn(threadId: string, persistedMessages: UiMessage[]): void {
+    const current = liveCommandsByThreadId.value[threadId]
+    if (!current || current.length === 0) return
+    const persistedIds = new Set(persistedMessages.map((m) => m.id))
+    const next = current.filter((m) => !persistedIds.has(m.id))
+    if (next.length === current.length) return
+    if (next.length === 0) {
+      liveCommandsByThreadId.value = omitKey(liveCommandsByThreadId.value, threadId)
+    } else {
+      liveCommandsByThreadId.value = { ...liveCommandsByThreadId.value, [threadId]: next }
+    }
+  }
+
+  function removeLiveFileChangesPersistedIn(threadId: string, persistedMessages: UiMessage[]): void {
+    const current = liveFileChangeMessagesByThreadId.value[threadId]
+    if (!current || current.length === 0) return
+    const persistedIds = new Set(persistedMessages.map((message) => message.id))
+    const persistedTurnIds = new Set(
+      persistedMessages
+        .filter((message) => message.messageType === 'fileChange' && typeof message.turnId === 'string' && message.turnId.length > 0)
+        .map((message) => message.turnId as string),
+    )
+    const persistedTurnIndices = new Set(
+      persistedMessages
+        .filter((message) => message.messageType === 'fileChange' && typeof message.turnIndex === 'number')
+        .map((message) => message.turnIndex as number),
+    )
+    const next = current.filter((message) => (
+      !persistedIds.has(message.id)
+      && !(message.turnId && persistedTurnIds.has(message.turnId))
+      && !(typeof message.turnIndex === 'number' && persistedTurnIndices.has(message.turnIndex))
+    ))
+    if (next.length === current.length) return
+    if (next.length === 0) {
+      liveFileChangeMessagesByThreadId.value = omitKey(liveFileChangeMessagesByThreadId.value, threadId)
+    } else {
+      liveFileChangeMessagesByThreadId.value = { ...liveFileChangeMessagesByThreadId.value, [threadId]: next }
+    }
+  }
+
+  function applyRealtimeUpdates(notification: RpcNotification): void {
+    if (handleServerRequestNotification(notification)) {
+      return
+    }
+
+    if (notification.method === 'account/rateLimits/updated') {
+      scheduleRateLimitRefresh()
+    }
+
+    if (notification.method === 'thread/name/updated') {
+      const params = asRecord(notification.params)
+      const threadId = readString(params?.threadId)
+      const threadName = readString(params?.threadName)
+      if (threadId && threadName) {
+        threadTitleById.value = { ...threadTitleById.value, [threadId]: threadName }
+        applyThreadFlags()
+        void persistThreadTitle(threadId, threadName)
+      }
+    }
+
+    if (notification.method === 'account/rateLimits/updated') {
+      setCodexRateLimit(pickCodexRateLimitSnapshot(notification.params))
+      return
+    }
+
+    const tokenUsageUpdate = readThreadTokenUsageUpdate(notification)
+    if (tokenUsageUpdate) {
+      setThreadTokenUsage(tokenUsageUpdate.threadId, tokenUsageUpdate.usage)
+      return
+    }
+
+    const turnActivity = readTurnActivity(notification)
+    if (turnActivity) {
+      setTurnActivityForThread(turnActivity.threadId, turnActivity.activity)
+    }
+
+    const notificationThreadId = extractThreadIdFromNotification(notification)
+    const notificationErrorState = readNotificationErrorState(notification)
+    if (!notificationErrorState && notificationThreadId) {
+      clearTransientTurnErrorForThread(notificationThreadId)
+    }
+
+    const startedTurn = readTurnStartedInfo(notification)
+    if (startedTurn) {
+      pendingTurnStartsById.set(startedTurn.turnId, startedTurn)
+      setTurnIndexForThread(startedTurn.threadId, startedTurn.turnId, inferNextTurnIndex(startedTurn.threadId))
+      activeTurnIdByThreadId.value = {
+        ...activeTurnIdByThreadId.value,
+        [startedTurn.threadId]: startedTurn.turnId,
+      }
+      maybeUnblockInterruptForActiveTurn(startedTurn.threadId, startedTurn.turnId)
+      clearLivePlansForThread(startedTurn.threadId)
+      clearLiveFileChangesForThread(startedTurn.threadId)
+      setTurnSummaryForThread(startedTurn.threadId, null)
+      setTurnErrorForThread(startedTurn.threadId, null)
+      setThreadInProgress(startedTurn.threadId, true)
+      scheduleQueueStateRefresh(startedTurn.threadId)
+      if (eventUnreadByThreadId.value[startedTurn.threadId]) {
+        eventUnreadByThreadId.value = omitKey(eventUnreadByThreadId.value, startedTurn.threadId)
+      }
+    }
+
+    const completedTurn = readTurnCompletedInfo(notification)
+    const turnErrorMessage = readTurnErrorMessage(notification)
+    const completedThreadId = completedTurn?.threadId ?? extractThreadIdFromNotification(notification)
+    const completedThreadModelId = completedThreadId ? readModelIdForThread(completedThreadId) : ''
+    const shouldRetryWithFallback =
+      Boolean(completedThreadId) &&
+      Boolean(turnErrorMessage) &&
+      completedThreadModelId !== MODEL_FALLBACK_ID &&
+      isUnsupportedChatGptModelError(new Error(turnErrorMessage))
+    if (completedTurn) {
+      const pendingTurnRequest = pendingTurnRequestByThreadId.value[completedTurn.threadId]
+      const startedTurnState = pendingTurnStartsById.get(completedTurn.turnId)
+      if (startedTurnState) {
+        pendingTurnStartsById.delete(completedTurn.turnId)
+      }
+
+      const rawDurationMs =
+        readNumber(asRecord(notification.params)?.durationMs) ??
+        readNumber(asRecord(asRecord(notification.params)?.turn)?.durationMs) ??
+        (typeof completedTurn.startedAtMs === 'number'
+          ? completedTurn.completedAtMs - completedTurn.startedAtMs
+          : null) ??
+        (startedTurnState ? completedTurn.completedAtMs - startedTurnState.startedAtMs : null)
+
+      const durationMs = typeof rawDurationMs === 'number' ? Math.max(0, rawDurationMs) : 0
+      setTurnSummaryForThread(completedTurn.threadId, {
+        turnId: completedTurn.turnId,
+        durationMs,
+      })
+      if (activeTurnIdByThreadId.value[completedTurn.threadId]) {
+        activeTurnIdByThreadId.value = omitKey(activeTurnIdByThreadId.value, completedTurn.threadId)
+      }
+      setThreadInProgress(completedTurn.threadId, false)
+      setTurnActivityForThread(completedTurn.threadId, null)
+      markThreadUnreadByEvent(completedTurn.threadId)
+      if (!shouldRetryWithFallback) {
+        clearPendingTurnRequest(completedTurn.threadId)
+        scheduleQueueStateRefresh(completedTurn.threadId)
+      }
+    }
+
+    if (turnErrorMessage) {
+      const failedThreadId = completedTurn?.threadId || extractThreadIdFromNotification(notification)
+      if (failedThreadId) {
+        setTurnErrorForThread(failedThreadId, turnErrorMessage)
+      }
+      error.value = turnErrorMessage
+      if (failedThreadId && shouldRetryWithFallback) {
+        void retryPendingTurnWithFallback(failedThreadId)
+      }
+    } else if (completedTurn) {
+      setTurnErrorForThread(completedTurn.threadId, null)
+    }
+
+    if (notificationErrorState) {
+      const errorThreadId = notificationThreadId
+      const errorThreadModelId = errorThreadId ? readModelIdForThread(errorThreadId) : selectedModelId.value.trim()
+      if (errorThreadId) {
+        setTurnErrorForThread(errorThreadId, notificationErrorState.message, {
+          transient: notificationErrorState.transient,
+        })
+      }
+      error.value = notificationErrorState.message
+      if (errorThreadModelId !== MODEL_FALLBACK_ID && isUnsupportedChatGptModelError(new Error(notificationErrorState.message))) {
+        if (errorThreadId) {
+          void retryPendingTurnWithFallback(errorThreadId)
+        } else {
+          void applyFallbackModelSelection()
+        }
+      }
+    }
+
+    const planUpdate = readPlanUpdate(notification)
+    if (planUpdate) {
+      upsertLivePlanMessage(planUpdate.threadId, planUpdate.message)
+      setTurnActivityForThread(planUpdate.threadId, {
+        label: 'Planning',
+        details: planUpdate.message.plan?.steps.map((step) => step.step).slice(0, 2) ?? [],
+      })
+    }
+
+    const planDelta = readPlanDelta(notification)
+    if (planDelta) {
+      upsertLivePlanMessage(planDelta.threadId, planDelta.message)
+      setTurnActivityForThread(planDelta.threadId, {
+        label: 'Planning',
+        details: [],
+      })
+    }
+
+    if (!notificationThreadId || notificationThreadId !== selectedThreadId.value) return
+
+    const startedAgentMessageId = readAgentMessageStartedId(notification)
+    if (startedAgentMessageId) {
+      activeReasoningItemId = ''
+    }
+
+    const liveAgentMessageDelta = readAgentMessageDelta(notification)
+    if (liveAgentMessageDelta) {
+      const existing = (liveAgentMessagesByThreadId.value[notificationThreadId] ?? [])
+        .find((message) => message.id === liveAgentMessageDelta.messageId)
+      const nextText = `${existing?.text ?? ''}${liveAgentMessageDelta.delta}`
+      upsertLiveAgentMessage(notificationThreadId, {
+        id: liveAgentMessageDelta.messageId,
+        role: 'assistant',
+        text: nextText,
+        messageType: 'agentMessage.live',
+      })
+    }
+
+    const completedAgentMessage = readAgentMessageCompleted(notification)
+    if (completedAgentMessage) {
+      upsertLiveAgentMessage(notificationThreadId, completedAgentMessage)
+    }
+
+    const completedImageView = readCompletedImageView(notification)
+    if (completedImageView) {
+      upsertLiveAgentMessage(notificationThreadId, completedImageView)
+
+    }
+
+    const startedReasoningItemId = readReasoningStartedItemId(notification)
+    if (startedReasoningItemId) {
+      activeReasoningItemId = startedReasoningItemId
+    }
+
+    const liveReasoningDelta = readReasoningDelta(notification)
+    if (liveReasoningDelta) {
+      appendLiveReasoningText(notificationThreadId, liveReasoningDelta.delta)
+    }
+
+    const sectionBreakMessageId = readReasoningSectionBreakMessageId(notification)
+    if (sectionBreakMessageId) {
+      const current = liveReasoningTextByThreadId.value[notificationThreadId] ?? ''
+      if (current.trim().length > 0 && !current.endsWith('\n\n')) {
+        setLiveReasoningText(notificationThreadId, `${current}\n\n`)
+      }
+    }
+
+    const completedReasoningMessageId = readReasoningCompletedId(notification)
+    if (completedReasoningMessageId) {
+      if (completedReasoningMessageId === liveReasoningMessageId(activeReasoningItemId)) {
+        activeReasoningItemId = ''
+      }
+    }
+
+    const commandStarted = readCommandExecutionStarted(notification)
+    if (commandStarted) {
+      upsertLiveCommand(notificationThreadId, commandStarted)
+      setTurnActivityForThread(notificationThreadId, { label: 'Running command', details: [commandStarted.commandExecution?.command ?? ''] })
+    }
+
+    const commandDelta = readCommandOutputDelta(notification)
+    if (commandDelta) {
+      const current = (liveCommandsByThreadId.value[notificationThreadId] ?? []).find((m) => m.id === commandDelta.itemId)
+      if (current?.commandExecution) {
+        upsertLiveCommand(notificationThreadId, {
+          ...current,
+          commandExecution: { ...current.commandExecution, aggregatedOutput: `${current.commandExecution.aggregatedOutput}${commandDelta.delta}` },
+        })
+      }
+    }
+
+    const commandCompleted = readCommandExecutionCompleted(notification)
+    if (commandCompleted) {
+      upsertLiveCommand(notificationThreadId, commandCompleted)
+    }
+
+    const completedFileChange = readCompletedFileChange(notification)
+    if (completedFileChange) {
+      upsertLiveFileChangeMessage(notificationThreadId, completedFileChange)
+    }
+
+    if (isAgentContentEvent(notification)) {
+      activeReasoningItemId = ''
+      clearLiveReasoningForThread(notificationThreadId)
+    }
+
+    if (notification.method === 'turn/completed') {
+      activeReasoningItemId = ''
+      shouldAutoScrollOnNextAgentEvent = false
+      clearLiveReasoningForThread(notificationThreadId)
+      if (liveCommandsByThreadId.value[notificationThreadId]) {
+        liveCommandsByThreadId.value = omitKey(liveCommandsByThreadId.value, notificationThreadId)
+      }
+      const completedThreadId = extractThreadIdFromNotification(notification)
+      if (completedThreadId) {
+        setThreadInProgress(completedThreadId, false)
+        setTurnActivityForThread(completedThreadId, null)
+        markThreadUnreadByEvent(completedThreadId)
+        if (!shouldRetryWithFallback) {
+          clearPendingTurnRequest(completedThreadId)
+          scheduleQueueStateRefresh(completedThreadId)
+        }
+      }
+    }
+
+  }
+
+  function queueEventDrivenSync(notification: RpcNotification): void {
+    if (notification.method === 'thread/tokenUsage/updated') return
+
+    const method = notification.method
+    const shouldRefreshMessages =
+      method === 'turn/started' ||
+      method === 'turn/completed' ||
+      method === 'error'
+    const shouldRefreshThreads =
+      method.startsWith('thread/') ||
+      method === 'turn/completed'
+
+    if (!shouldRefreshMessages && !shouldRefreshThreads) return
+
+    const threadId = extractThreadIdFromNotification(notification)
+    if (threadId && shouldRefreshMessages) {
+      pendingThreadMessageRefresh.add(threadId)
+    }
+
+    if (shouldRefreshThreads) {
+      pendingThreadsRefresh = true
+    }
+
+    if (eventSyncTimer !== null || typeof window === 'undefined') return
+    eventSyncTimer = window.setTimeout(() => {
+      eventSyncTimer = null
+      void syncFromNotifications()
+    }, EVENT_SYNC_DEBOUNCE_MS)
+  }
+
+  async function hydrateWorkspaceRootsStateIfNeeded(
+    groups: UiProjectGroup[],
+    rootsState: WorkspaceRootsState | null,
+  ): Promise<void> {
+    if (hasHydratedWorkspaceRootsState) return
+    hasHydratedWorkspaceRootsState = true
+
+    try {
+      if (!rootsState) return
+      const hydratedOrder: string[] = []
+      for (const rootPath of getWorkspaceProjectOrderPaths(rootsState)) {
+        const projectName = toProjectNameFromWorkspaceRoot(rootPath)
+        if (hydratedOrder.includes(projectName)) continue
+        hydratedOrder.push(projectName)
+      }
+
+      if (hydratedOrder.length > 0) {
+        const mergedOrder = rootsState.projectOrder.length > 0
+          ? mergeProjectOrder(hydratedOrder, groups)
+          : mergeProjectOrder(projectOrder.value, groups)
+        if (!areStringArraysEqual(projectOrder.value, mergedOrder)) {
+          projectOrder.value = mergedOrder
+        }
+      }
+
+      if (Object.keys(rootsState.labels).length > 0 || (rootsState.remoteProjects ?? []).length > 0) {
+        const nextLabels = { ...projectDisplayNameById.value }
+        let changed = false
+        for (const [rootPath, label] of Object.entries(rootsState.labels)) {
+          const normalizedRootPath = normalizePathForUi(rootPath).trim()
+          const projectNames = [toProjectNameFromWorkspaceRoot(rootPath)]
+          if (normalizedRootPath) projectNames.push(normalizedRootPath)
+          for (const projectName of projectNames) {
+            if (nextLabels[projectName] === label) continue
+            nextLabels[projectName] = label
+            changed = true
+          }
+        }
+        for (const rootPath of rootsState.order) {
+          const leafName = toProjectNameFromWorkspaceRoot(rootPath)
+          const parentLeafName = toProjectName(getPathParent(rootPath))
+          if (!parentLeafName.startsWith('.') || parentLeafName === leafName) continue
+          const displayName = `${leafName} ${parentLeafName}`
+          if (nextLabels[leafName] !== undefined || nextLabels[leafName] === displayName) continue
+          nextLabels[leafName] = displayName
+          changed = true
+        }
+        for (const remoteProject of rootsState.remoteProjects ?? []) {
+          const label = getRemoteProjectDisplayName(remoteProject)
+          if (nextLabels[remoteProject.id] === label) continue
+          nextLabels[remoteProject.id] = label
+          changed = true
+        }
+        if (changed) {
+          projectDisplayNameById.value = nextLabels
+        }
+      }
+    } catch {
+      // Keep local storage fallback when global state is unavailable.
+    }
+  }
+
+  async function loadThreadTitleCacheIfNeeded(): Promise<void> {
+    if (Object.keys(threadTitleById.value).length > 0) return
+    try {
+      const cache = await getThreadTitleCache()
+      if (Object.keys(cache.titles).length > 0) {
+        threadTitleById.value = cache.titles
+      }
+    } catch {
+      // Title cache is optional; keep UI functional.
+    }
+  }
+
+  async function loadWorkspaceRootsStateForThreadList(): Promise<WorkspaceRootsState | null> {
+    try {
+      return await getWorkspaceRootsState()
+    } catch {
+      return null
+    }
+  }
+
+  async function requestThreadTitleGeneration(threadId: string, prompt: string, cwd: string | null): Promise<void> {
+    if (threadTitleById.value[threadId]) return
+    const trimmed = prompt.trim()
+    if (!trimmed) return
+    const truncated = trimmed.length > 300 ? trimmed.slice(0, 300) : trimmed
+    try {
+      const title = await generateThreadTitle(truncated, cwd)
+      if (!title || threadTitleById.value[threadId]) return
+      threadTitleById.value = { ...threadTitleById.value, [threadId]: title }
+      applyThreadFlags()
+      void persistThreadTitle(threadId, title)
+    } catch {
+      // Title generation is best-effort.
+    }
+  }
+
+  function filterGroupsByWorkspaceRoots(
+    groups: UiProjectGroup[],
+    rootsState: WorkspaceRootsState | null,
+  ): UiProjectGroup[] {
+    const duplicateLeafNames = collectDuplicateProjectLeafNames(groups, rootsState)
+    const disambiguatedGroups = disambiguateProjectGroupsByCwd(groups, rootsState)
+    const groupsWithWorkspaceRoots = addWorkspaceRootPlaceholderGroups(disambiguatedGroups, rootsState, duplicateLeafNames)
+    if (!rootsState || (rootsState.order.length === 0 && (rootsState.remoteProjects ?? []).length === 0)) return groupsWithWorkspaceRoots
+    const allowedProjectNames = new Set<string>()
+    for (const projectName of getWorkspaceProjectOrderNames(rootsState, duplicateLeafNames)) {
+      allowedProjectNames.add(projectName)
+    }
+    const filteredGroups = groupsWithWorkspaceRoots.filter((group) => {
+      if (allowedProjectNames.has(group.projectName)) return true
+      return isProjectlessGroup(group)
+    })
+    return orderGroupsByWorkspaceProjectOrder(filteredGroups, rootsState, duplicateLeafNames)
+  }
+
+  function applyThreadGroups(groups: UiProjectGroup[], rootsState: WorkspaceRootsState | null): void {
+    const visibleGroups = filterGroupsByWorkspaceRoots(groups, rootsState)
+    const hasWorkspaceRootsState = Boolean(
+      rootsState && (rootsState.order.length > 0 || rootsState.projectOrder.length > 0 || (rootsState.remoteProjects ?? []).length > 0),
+    )
+
+    const nextProjectOrder = rootsState?.projectOrder.length
+      ? mergeProjectOrder(
+        getWorkspaceProjectOrderNames(rootsState, collectDuplicateProjectLeafNames(groups, rootsState)),
+        visibleGroups,
+      )
+      : mergeProjectOrder(projectOrder.value, visibleGroups)
+    if (!areStringArraysEqual(projectOrder.value, nextProjectOrder)) {
+      projectOrder.value = nextProjectOrder
+      if (!hasWorkspaceRootsState) {
+        saveProjectOrder(projectOrder.value)
+      }
+    }
+
+    const orderedGroups = orderGroupsByProjectOrder(visibleGroups, projectOrder.value)
+    markServerListedThreads(new Set(flattenThreads(orderedGroups).map((thread) => thread.id)))
+    const mergedWithInProgress = mergeIncomingWithLocalInProgressThreads(
+      sourceGroups.value,
+      orderedGroups,
+      inProgressById.value,
+    )
+    sourceGroups.value = mergeThreadGroups(sourceGroups.value, mergedWithInProgress)
+    inProgressById.value = pruneThreadStateMap(
+      inProgressById.value,
+      new Set(flattenThreads(sourceGroups.value).map((thread) => thread.id)),
+    )
+    applyThreadFlags()
+  }
+
+  function normalizeQueueStateForPersistence(state: Record<string, QueuedMessage[]>): ThreadQueueState {
+    const next: ThreadQueueState = {}
+    for (const [threadId, queue] of Object.entries(state)) {
+      const normalizedThreadId = threadId.trim()
+      if (!normalizedThreadId || queue.length === 0) continue
+      next[normalizedThreadId] = queue.map((message) => ({
+        id: message.id,
+        text: message.text,
+        imageUrls: [...message.imageUrls],
+        skills: message.skills.map((skill) => ({ name: skill.name, path: skill.path })),
+        fileAttachments: message.fileAttachments.map((attachment) => ({
+          label: attachment.label,
+          path: attachment.path,
+          fsPath: attachment.fsPath,
+        })),
+        collaborationMode: message.collaborationMode,
+      }))
+    }
+    return next
+  }
+
+  function persistQueueState(): void {
+    void setThreadQueueState(normalizeQueueStateForPersistence(queuedMessagesByThreadId.value)).catch(() => {
+      // Queue persistence is best-effort; keep the current in-memory queue usable.
+    })
+  }
+
+  async function loadPersistedQueueStateIfNeeded(): Promise<void> {
+    if (hasLoadedPersistedQueueState) return
+    hasLoadedPersistedQueueState = true
+    try {
+      queuedMessagesByThreadId.value = await getThreadQueueState()
+    } catch {
+      // Backend queue state is optional during startup.
+    }
+  }
+
+  function removeArchivedThreadFromLoadedLists(threadId: string): void {
+    loadedThreadListGroups = removeThreadFromGroups(loadedThreadListGroups, threadId)
+    sourceGroups.value = removeThreadFromGroups(sourceGroups.value, threadId)
+    inProgressById.value = omitKey(inProgressById.value, threadId)
+    applyThreadFlags()
+  }
+
+  function mergeThreadGroupPages(previous: UiProjectGroup[], incoming: UiProjectGroup[]): UiProjectGroup[] {
+    if (previous.length === 0) return incoming
+    if (incoming.length === 0) return previous
+
+    const threadById = new Map<string, UiThread>()
+    for (const thread of flattenThreads(previous)) {
+      threadById.set(thread.id, thread)
+    }
+    for (const thread of flattenThreads(incoming)) {
+      threadById.set(thread.id, thread)
+    }
+    const groupsByProject = new Map<string, UiThread[]>()
+    for (const thread of threadById.values()) {
+      const existing = groupsByProject.get(thread.projectName)
+      if (existing) existing.push(thread)
+      else groupsByProject.set(thread.projectName, [thread])
+    }
+
+    return Array.from(groupsByProject.entries())
+      .map(([projectName, threads]) => ({
+        projectName,
+        threads: threads.sort(
+          (first, second) => new Date(second.updatedAtIso).getTime() - new Date(first.updatedAtIso).getTime(),
+        ),
+      }))
+      .sort((first, second) => {
+        const firstUpdated = new Date(first.threads[0]?.updatedAtIso ?? 0).getTime()
+        const secondUpdated = new Date(second.threads[0]?.updatedAtIso ?? 0).getTime()
+        return secondUpdated - firstUpdated
+      })
+  }
+
+  function hasActiveInProgressThreads(): boolean {
+    return Object.values(inProgressById.value).some((value) => value === true)
+  }
+
+  function scheduleRemainingThreadPages(rootsState: WorkspaceRootsState | null = loadedThreadListRootsState): void {
+    if (!threadListNextCursor || isLoadingRemainingThreadPages || hasActiveInProgressThreads()) return
+
+    loadedThreadListRootsState = rootsState
+
+    if (typeof window === 'undefined') {
+      void loadRemainingThreadPages(rootsState)
+      return
+    }
+
+    if (threadListBackgroundTimer !== null) {
+      window.clearTimeout(threadListBackgroundTimer)
+    }
+
+    threadListBackgroundTimer = window.setTimeout(() => {
+      threadListBackgroundTimer = null
+      if (!threadListNextCursor || hasActiveInProgressThreads()) return
+      void loadRemainingThreadPages(loadedThreadListRootsState)
+    }, BACKGROUND_THREAD_PAGINATION_DELAY_MS)
+  }
+
+  async function loadRemainingThreadPages(rootsState: WorkspaceRootsState | null): Promise<void> {
+    if (isLoadingRemainingThreadPages || !threadListNextCursor || hasActiveInProgressThreads()) return
+    isLoadingRemainingThreadPages = true
+
+    try {
+      const page = await getThreadGroupsPage(threadListNextCursor, getBackgroundThreadListLimit())
+      threadListNextCursor = page.nextCursor
+      hasLoadedAllThreadPages = page.nextCursor === null
+      isThreadListFullyLoaded.value = hasLoadedAllThreadPages
+      loadedThreadListGroups = mergeThreadGroupPages(loadedThreadListGroups, page.groups)
+      applyThreadGroups(loadedThreadListGroups, rootsState)
+    } catch {
+      // Keep the first page usable; a later refresh can retry remaining pages.
+    } finally {
+      isLoadingRemainingThreadPages = false
+      if (threadListNextCursor && !hasActiveInProgressThreads()) {
+        scheduleRemainingThreadPages(rootsState)
+      }
+    }
+  }
+
+  async function loadThreads() {
+    if (loadThreadsPromise) {
+      await loadThreadsPromise
+      return
+    }
+
+    loadThreadsPromise = (async () => {
+    if (!hasLoadedThreads.value) {
+      isLoadingThreads.value = true
+    }
+
+    try {
+      const [page, rootsState] = await Promise.all([
+        getThreadGroupsPage(),
+        loadWorkspaceRootsStateForThreadList(),
+        loadThreadTitleCacheIfNeeded(),
+      ])
+      loadedThreadListRootsState = rootsState
+      const groups = page.groups
+      loadedThreadListGroups = hasLoadedThreads.value
+        ? mergeThreadGroupPages(loadedThreadListGroups, groups)
+        : groups
+      threadListNextCursor = hasLoadedThreads.value && !hasLoadedAllThreadPages
+        ? threadListNextCursor
+        : page.nextCursor
+      hasLoadedAllThreadPages = page.nextCursor === null
+      isThreadListFullyLoaded.value = hasLoadedAllThreadPages
+      await hydrateWorkspaceRootsStateIfNeeded(groups, rootsState)
+
+      applyThreadGroups(loadedThreadListGroups, rootsState)
+      hasLoadedThreads.value = true
+      if (!hasLoadedAllThreadPages) {
+        scheduleRemainingThreadPages(rootsState)
+      }
+
+      const flatThreads = flattenThreads(projectGroups.value)
+      pruneThreadScopedState(flatThreads)
+
+      const currentExists = flatThreads.some((thread) => thread.id === selectedThreadId.value)
+
+      if (!currentExists && !selectedThreadId.value) {
+        setSelectedThreadId(flatThreads[0]?.id ?? '')
+      }
+    } finally {
+      isLoadingThreads.value = false
+    }
+    })().finally(() => {
+      loadThreadsPromise = null
+    })
+
+    await loadThreadsPromise
+  }
+
+  async function loadMessages(threadId: string, options: { silent?: boolean } = {}) {
+    if (!threadId) {
+      return
+    }
+
+    const existingLoad = loadMessagePromiseByThreadId.get(threadId)
+    if (existingLoad) {
+      await existingLoad
+      return
+    }
+
+    const alreadyLoaded = loadedMessagesByThreadId.value[threadId] === true
+    const shouldShowLoading = options.silent !== true && !alreadyLoaded
+    if (shouldShowLoading) {
+      isLoadingMessages.value = true
+    }
+
+    const loadPromise = (async () => {
+      try {
+      const version = currentThreadVersion(threadId)
+      const loadedVersion = loadedVersionByThreadId.value[threadId] ?? ''
+      const loadedRecently =
+        Date.now() - (lastMessageLoadAtByThreadId.get(threadId) ?? 0) < RECENT_THREAD_MESSAGE_LOAD_REUSE_MS
+      const canReuseLoadedMessages =
+        alreadyLoaded &&
+        (
+          loadedRecently ||
+          (
+            (version.length === 0 || loadedVersion === version) &&
+            inProgressById.value[threadId] !== true
+          )
+        )
+
+      if (canReuseLoadedMessages) {
+        markThreadAsRead(threadId)
+        return
+      }
+
+      const needsResume = resumedThreadById.value[threadId] !== true
+      const resumedThread = needsResume ? await resumeThread(threadId) : null
+      const detail = resumedThread ?? await getThreadDetail(threadId)
+
+      if (resumedThread) {
+        setThreadModelId(threadId, resumedThread.model)
+        resumedThreadById.value = {
+          ...resumedThreadById.value,
+          [threadId]: true,
+        }
+      }
+
+      const { messages: nextMessages, inProgress, activeTurnId, turnIndexByTurnId } = detail
+      hasMoreOlderMessagesByThreadId.value = {
+        ...hasMoreOlderMessagesByThreadId.value,
+        [threadId]: detail.hasMoreOlder === true,
+      }
+      markThreadMessagesPersisted(threadId, nextMessages)
+      replaceTurnIndexLookupForThread(threadId, turnIndexByTurnId)
+      rebindLiveFileChangeTurnIndices(threadId)
+      const previousPersisted = persistedMessagesByThreadId.value[threadId] ?? []
+      const mergedMessages = mergeMessages(previousPersisted, nextMessages, {
+        preserveMissing: options.silent === true,
+      })
+      setPersistedMessagesForThread(threadId, mergedMessages)
+
+      const previousLiveAgent = liveAgentMessagesByThreadId.value[threadId] ?? []
+      if (inProgress) {
+        const nextLiveAgent = removeRedundantLiveAgentMessages(previousLiveAgent, nextMessages)
+        setLiveAgentMessagesForThread(threadId, nextLiveAgent)
+      } else {
+        clearLiveAgentMessagesForThread(threadId)
+      }
+      removeLiveCommandsPersistedIn(threadId, nextMessages)
+      removeLiveFileChangesPersistedIn(threadId, nextMessages)
+
+      loadedMessagesByThreadId.value = {
+        ...loadedMessagesByThreadId.value,
+        [threadId]: true,
+      }
+      lastMessageLoadAtByThreadId.set(threadId, Date.now())
+
+      if (version) {
+        loadedVersionByThreadId.value = {
+          ...loadedVersionByThreadId.value,
+          [threadId]: version,
+        }
+      }
+      setThreadInProgress(threadId, inProgress)
+      if (activeTurnId) {
+        activeTurnIdByThreadId.value = {
+          ...activeTurnIdByThreadId.value,
+          [threadId]: activeTurnId,
+        }
+      } else if (activeTurnIdByThreadId.value[threadId]) {
+        activeTurnIdByThreadId.value = omitKey(activeTurnIdByThreadId.value, threadId)
+      }
+      if (!inProgress) {
+        clearCompletedTurnLiveState(threadId)
+      }
+      markThreadAsRead(threadId)
+      } finally {
+      if (shouldShowLoading) {
+        isLoadingMessages.value = false
+      }
+      }
+    })().finally(() => {
+      loadMessagePromiseByThreadId.delete(threadId)
+    })
+
+    loadMessagePromiseByThreadId.set(threadId, loadPromise)
+    await loadPromise
+  }
+
+  async function loadOlderMessages(threadId: string = selectedThreadId.value): Promise<void> {
+    if (!threadId) return
+    if (loadingOlderMessagesByThreadId.value[threadId] === true) return
+    if (hasMoreOlderMessagesByThreadId.value[threadId] !== true) return
+
+    const beforeTurnId = getFirstPersistedTurnId(threadId)
+    if (!beforeTurnId) {
+      hasMoreOlderMessagesByThreadId.value = {
+        ...hasMoreOlderMessagesByThreadId.value,
+        [threadId]: false,
+      }
+      return
+    }
+
+    loadingOlderMessagesByThreadId.value = {
+      ...loadingOlderMessagesByThreadId.value,
+      [threadId]: true,
+    }
+
+    try {
+      const page = await getOlderThreadMessages(threadId, beforeTurnId)
+      const previousPersisted = persistedMessagesByThreadId.value[threadId] ?? []
+      const mergedMessages = mergeMessages(page.messages, previousPersisted, { preserveMissing: true })
+      setPersistedMessagesForThread(threadId, mergedMessages)
+      replaceTurnIndexLookupForThread(threadId, {
+        ...(turnIndexByTurnIdByThreadId.value[threadId] ?? {}),
+        ...page.turnIndexByTurnId,
+      })
+      rebindLiveFileChangeTurnIndices(threadId)
+      hasMoreOlderMessagesByThreadId.value = {
+        ...hasMoreOlderMessagesByThreadId.value,
+        [threadId]: page.hasMoreOlder,
+      }
+    } catch (loadError) {
+      error.value = loadError instanceof Error ? loadError.message : 'Failed to load earlier messages'
+      throw loadError
+    } finally {
+      loadingOlderMessagesByThreadId.value = {
+        ...loadingOlderMessagesByThreadId.value,
+        [threadId]: false,
+      }
+    }
+  }
+
+  async function ensureThreadMessagesLoaded(threadId: string, options: { silent?: boolean } = {}): Promise<void> {
+    if (!threadId) return
+    if (loadedMessagesByThreadId.value[threadId] === true) return
+    await loadMessages(threadId, options)
+  }
+
+  async function refreshSkills(): Promise<void> {
+    if (refreshSkillsPromise) {
+      await refreshSkillsPromise
+      return
+    }
+
+    refreshSkillsPromise = (async () => {
+      try {
+        const selectedCwd = selectedThread.value?.cwd?.trim() ?? ''
+        installedSkills.value = await getSkillsList(selectedCwd ? [selectedCwd] : undefined)
+      } catch {
+        // keep previous skills on failure
+      } finally {
+        refreshSkillsPromise = null
+      }
+    })()
+
+    await refreshSkillsPromise
+  }
+
+  async function refreshAncillaryState(
+    options: { providerChanged?: boolean; includeProviderModels?: boolean } = {},
+  ): Promise<void> {
+    await Promise.allSettled([
+      refreshModelPreferences({
+        providerChanged: options.providerChanged,
+        includeProviderModels: options.includeProviderModels,
+      }),
+      refreshRateLimits(),
+      refreshCollaborationModes(),
+      refreshSkills(),
+    ])
+  }
+
+  function scheduleAncillaryStateRefresh(
+    options: { providerChanged?: boolean; includeProviderModels?: boolean } = {},
+  ): void {
+    const run = () => {
+      void refreshAncillaryState(options)
+    }
+
+    if (typeof window === 'undefined') {
+      run()
+      return
+    }
+
+    window.setTimeout(run, 0)
+  }
+
+  async function refreshAll(
+    options: { includeSelectedThreadMessages?: boolean; awaitAncillaryRefreshes?: boolean; providerChanged?: boolean } = {},
+  ) {
+    error.value = ''
+    const includeSelectedThreadMessages = options.includeSelectedThreadMessages !== false
+    const awaitAncillaryRefreshes = options.awaitAncillaryRefreshes === true
+
+    try {
+      await loadPersistedQueueStateIfNeeded()
+      await loadThreads()
+      if (includeSelectedThreadMessages) {
+        await loadMessages(selectedThreadId.value)
+      }
+      if (awaitAncillaryRefreshes) {
+        await refreshAncillaryState({
+          providerChanged: options.providerChanged,
+          includeProviderModels: options.providerChanged === true || awaitAncillaryRefreshes,
+        })
+      } else {
+        scheduleAncillaryStateRefresh({
+          providerChanged: options.providerChanged,
+          includeProviderModels: false,
+        })
+      }
+    } catch (unknownError) {
+      error.value = unknownError instanceof Error ? unknownError.message : 'Unknown application error'
+    }
+  }
+
+  async function selectThread(threadId: string) {
+    setSelectedThreadId(threadId)
+
+    try {
+      await loadMessages(threadId)
+      void refreshSkills()
+    } catch (unknownError) {
+      error.value = unknownError instanceof Error ? unknownError.message : 'Unknown application error'
+    }
+  }
+
+  async function archiveThreadById(threadId: string) {
+    const wasSelectedThread = selectedThreadId.value === threadId
+    const nextSelectedThreadId = wasSelectedThread
+      ? findAdjacentThreadId(flattenThreads(projectGroups.value), threadId)
+      : ''
+
+    if (wasSelectedThread) {
+      setSelectedThreadId(nextSelectedThreadId)
+      if (nextSelectedThreadId) {
+        void loadMessages(nextSelectedThreadId, { silent: true })
+      }
+    }
+
+    try {
+      await archiveThread(threadId)
+      removeArchivedThreadFromLoadedLists(threadId)
+      await loadThreads()
+
+      if (wasSelectedThread && nextSelectedThreadId && selectedThreadId.value === nextSelectedThreadId) {
+        await ensureThreadMessagesLoaded(nextSelectedThreadId, { silent: true })
+      }
+    } catch (unknownError) {
+      error.value = unknownError instanceof Error ? unknownError.message : 'Unknown application error'
+    }
+  }
+
+  async function renameThreadById(threadId: string, threadName: string) {
+    const normalizedName = threadName.trim()
+    if (!threadId || !normalizedName) return
+
+    try {
+      await renameThread(threadId, normalizedName)
+      threadTitleById.value = { ...threadTitleById.value, [threadId]: normalizedName }
+      applyThreadFlags()
+      void persistThreadTitle(threadId, normalizedName)
+    } catch (unknownError) {
+      error.value = unknownError instanceof Error ? unknownError.message : 'Unknown application error'
+    }
+  }
+
+  async function forkThreadById(threadId: string): Promise<string> {
+    const sourceThreadId = threadId.trim()
+    if (!sourceThreadId) return ''
+
+    const sourceThread = flattenThreads(sourceGroups.value).find((row) => row.id === sourceThreadId)
+    const sourceCwd = sourceThread?.cwd?.trim() ?? ''
+    const sourceTitle = sourceThread?.title?.trim() ?? 'Forked chat'
+    const selectedModel = readModelIdForThread(sourceThreadId)
+    error.value = ''
+
+    try {
+      const forkedThread = await forkThread(sourceThreadId, sourceCwd || undefined, selectedModel || undefined)
+      const nextThreadId = forkedThread.threadId.trim()
+      if (!nextThreadId) return ''
+
+      insertOptimisticThread(nextThreadId, sourceCwd, sourceTitle)
+      setThreadModelId(nextThreadId, forkedThread.model)
+      resumedThreadById.value = {
+        ...resumedThreadById.value,
+        [nextThreadId]: true,
+      }
+      setSelectedThreadId(nextThreadId)
+      await loadThreads()
+      await loadMessages(nextThreadId)
+      return nextThreadId
+    } catch (unknownError) {
+      error.value = unknownError instanceof Error ? unknownError.message : 'Unknown application error'
+      return ''
+    }
+  }
+
+  async function forkThreadFromTurn(threadId: string, turnIndex: number): Promise<string> {
+    const normalizedThreadId = threadId.trim()
+    if (!normalizedThreadId || !Number.isInteger(turnIndex) || turnIndex < 0) return ''
+
+    if (inProgressById.value[normalizedThreadId] === true) {
+      error.value = 'Finish the current turn before forking from a response.'
+      return ''
+    }
+
+    if (loadedMessagesByThreadId.value[normalizedThreadId] !== true) {
+      try {
+        await loadMessages(normalizedThreadId)
+      } catch (unknownError) {
+        error.value = unknownError instanceof Error ? unknownError.message : 'Unknown application error'
+        return ''
+      }
+    }
+
+    const sourceMessages = persistedMessagesByThreadId.value[normalizedThreadId] ?? []
+    let lastTurnIndex = -1
+    for (const message of sourceMessages) {
+      if (typeof message.turnIndex === 'number' && Number.isFinite(message.turnIndex)) {
+        lastTurnIndex = Math.max(lastTurnIndex, message.turnIndex)
+      }
+    }
+
+    if (lastTurnIndex >= 0 && turnIndex > lastTurnIndex) return ''
+
+    const sourceThread = flattenThreads(sourceGroups.value).find((row) => row.id === normalizedThreadId) ?? null
+
+    try {
+      error.value = ''
+      const forked = await forkThread(normalizedThreadId)
+      const forkedThreadId = forked.threadId.trim()
+      if (!forkedThreadId) return ''
+
+      const forkedCwd = forked.cwd.trim() || sourceThread?.cwd?.trim() || ''
+      const forkedThreadTitle = toForkedThreadTitle(sourceThread?.title || sourceThread?.preview || 'Untitled thread')
+      insertOptimisticThread(forkedThreadId, forkedCwd, forkedThreadTitle)
+      setThreadModelId(forkedThreadId, forked.model)
+      setPersistedMessagesForThread(forkedThreadId, forked.messages)
+      loadedMessagesByThreadId.value = {
+        ...loadedMessagesByThreadId.value,
+        [forkedThreadId]: true,
+      }
+      resumedThreadById.value = {
+        ...resumedThreadById.value,
+        [forkedThreadId]: true,
+      }
+      clearLivePlansForThread(forkedThreadId)
+      setLiveAgentMessagesForThread(forkedThreadId, [])
+      clearLiveReasoningForThread(forkedThreadId)
+      if (liveCommandsByThreadId.value[forkedThreadId]) {
+        liveCommandsByThreadId.value = omitKey(liveCommandsByThreadId.value, forkedThreadId)
+      }
+      setTurnSummaryForThread(forkedThreadId, null)
+      setTurnActivityForThread(forkedThreadId, null)
+      setTurnErrorForThread(forkedThreadId, null)
+      setThreadInProgress(forkedThreadId, false)
+
+      const turnsToRollback = lastTurnIndex - turnIndex
+      if (turnsToRollback > 0) {
+        const rolledBackMessages = await rollbackThread(forkedThreadId, turnsToRollback)
+        setPersistedMessagesForThread(forkedThreadId, rolledBackMessages)
+      }
+
+      await renameThreadById(forkedThreadId, forkedThreadTitle)
+      setSelectedThreadId(forkedThreadId)
+      void loadThreads().catch(() => {})
+      return forkedThreadId
+    } catch (unknownError) {
+      error.value = unknownError instanceof Error ? unknownError.message : 'Unknown application error'
+      return ''
+    }
+  }
+
+  async function maybeReplyToPendingUserInputRequest(
+    threadId: string,
+    text: string,
+    imageUrls: string[] = [],
+    skills: Array<{ name: string; path: string }> = [],
+    fileAttachments: FileAttachment[] = [],
+  ): Promise<boolean> {
+    if (!threadId || !text.trim()) return false
+    if (imageUrls.length > 0 || skills.length > 0 || fileAttachments.length > 0) return false
+
+    const requests = pendingServerRequestsByThreadId.value[threadId] ?? []
+    const userInputRequests = requests.filter((request) => request.method === 'item/tool/requestUserInput')
+    if (userInputRequests.length !== 1) return false
+
+    const [request] = userInputRequests
+    const questionIds = readToolRequestUserInputQuestionIds(request)
+    if (questionIds.length !== 1) return false
+
+    return respondToPendingServerRequest({
+      id: request.id,
+      result: {
+        answers: {
+          [questionIds[0]]: {
+            answers: [text.trim()],
+          },
+        },
+      },
+    })
+  }
+
+  async function sendMessageToSelectedThread(
+    text: string,
+    imageUrls: string[] = [],
+    skills: Array<{ name: string; path: string }> = [],
+    mode: 'steer' | 'queue' = 'steer',
+    fileAttachments: FileAttachment[] = [],
+    queueInsertIndex?: number,
+    collaborationModeOverride?: CollaborationModeKind,
+  ): Promise<void> {
+    if (isUpdatingSpeedMode.value) return
+
+    const threadId = selectedThreadId.value
+    const nextText = text.trim()
+    if (!threadId || (!nextText && imageUrls.length === 0 && fileAttachments.length === 0)) return
+
+    if (await maybeReplyToPendingUserInputRequest(threadId, nextText, imageUrls, skills, fileAttachments)) {
+      return
+    }
+
+    const isInProgress = inProgressById.value[threadId] === true
+
+    if (isInProgress && mode === 'queue') {
+      const queue = queuedMessagesByThreadId.value[threadId] ?? []
+      const id = `q-${Date.now()}-${Math.random().toString(36).slice(2, 8)}`
+      const nextQueue = [...queue]
+      const insertIndex = typeof queueInsertIndex === 'number'
+        ? Math.max(0, Math.min(queueInsertIndex, nextQueue.length))
+        : nextQueue.length
+      nextQueue.splice(insertIndex, 0, {
+        id,
+        text: nextText,
+        imageUrls,
+        skills,
+        fileAttachments,
+        collaborationMode: collaborationModeOverride === 'plan'
+          ? 'plan'
+          : collaborationModeOverride === 'default'
+            ? 'default'
+            : selectedCollaborationMode.value,
+      })
+      queuedMessagesByThreadId.value = {
+        ...queuedMessagesByThreadId.value,
+        [threadId]: nextQueue,
+      }
+      persistQueueState()
+      return
+    }
+
+    if (isInProgress) {
+      shouldAutoScrollOnNextAgentEvent = true
+      void startTurnForThread(
+        threadId,
+        nextText,
+        imageUrls,
+        skills,
+        fileAttachments,
+        collaborationModeOverride,
+      ).catch((unknownError) => {
+        const errorMessage = unknownError instanceof Error ? unknownError.message : 'Unknown application error'
+        setTurnErrorForThread(threadId, errorMessage)
+        error.value = errorMessage
+      })
+      return
+    }
+
+    error.value = ''
+    shouldAutoScrollOnNextAgentEvent = true
+    setTurnSummaryForThread(threadId, null)
+    setTurnActivityForThread(
+      threadId,
+      {
+        label: 'Thinking',
+        details: buildPendingTurnDetails(
+          readModelIdForThread(threadId),
+          selectedReasoningEffort.value,
+          collaborationModeOverride === 'plan'
+            ? 'plan'
+            : collaborationModeOverride === 'default'
+              ? 'default'
+              : selectedCollaborationMode.value,
+        ),
+      },
+    )
+    setTurnErrorForThread(threadId, null)
+    setThreadInProgress(threadId, true)
+
+    try {
+      await startTurnForThread(
+        threadId,
+        nextText,
+        imageUrls,
+        skills,
+        fileAttachments,
+        collaborationModeOverride,
+      )
+    } catch (unknownError) {
+      shouldAutoScrollOnNextAgentEvent = false
+      setThreadInProgress(threadId, false)
+      setTurnActivityForThread(threadId, null)
+      const errorMessage = unknownError instanceof Error ? unknownError.message : 'Unknown application error'
+      setTurnErrorForThread(threadId, errorMessage)
+      error.value = errorMessage
+      throw unknownError
+    }
+  }
+
+  async function sendMessageToNewThread(
+    text: string,
+    cwd: string,
+    imageUrls: string[] = [],
+    skills: Array<{ name: string; path: string }> = [],
+    fileAttachments: FileAttachment[] = [],
+  ): Promise<string> {
+    if (isUpdatingSpeedMode.value) return ''
+
+    const nextText = text.trim()
+    const targetCwd = cwd.trim()
+    const selectedModel = readModelIdForThread(NEW_THREAD_COLLABORATION_MODE_CONTEXT).trim()
+    const selectedMode = selectedCollaborationMode.value
+    if (!nextText && imageUrls.length === 0 && fileAttachments.length === 0) return ''
+
+    isSendingMessage.value = true
+    error.value = ''
+    let threadId = ''
+
+    try {
+      try {
+        const startedThread = await startThread(targetCwd || undefined, selectedModel || undefined)
+        threadId = startedThread.threadId
+        setThreadModelId(threadId, startedThread.model)
+        setSelectedCollaborationModeForThread(threadId, selectedMode)
+      } catch (unknownError) {
+        if (selectedModel && selectedModel !== MODEL_FALLBACK_ID && isUnsupportedChatGptModelError(unknownError)) {
+          await applyFallbackModelSelection()
+          const fallbackThread = await startThread(targetCwd || undefined, MODEL_FALLBACK_ID)
+          threadId = fallbackThread.threadId
+          setThreadModelId(threadId, fallbackThread.model)
+          setSelectedCollaborationModeForThread(threadId, selectedMode)
+        } else {
+          throw unknownError
+        }
+      }
+      if (!threadId) return ''
+
+      insertOptimisticThread(threadId, targetCwd, nextText || '[Image]')
+      blockInterruptUntilThreadIsPersisted(threadId)
+      resumedThreadById.value = {
+        ...resumedThreadById.value,
+        [threadId]: true,
+      }
+      setSelectedThreadId(threadId)
+      shouldAutoScrollOnNextAgentEvent = true
+      setTurnSummaryForThread(threadId, null)
+      setTurnActivityForThread(
+        threadId,
+        {
+          label: 'Thinking',
+          details: buildPendingTurnDetails(
+            readModelIdForThread(threadId),
+            selectedReasoningEffort.value,
+            selectedMode,
+          ),
+        },
+      )
+      setTurnErrorForThread(threadId, null)
+      setThreadInProgress(threadId, true)
+      const capturedThreadId = threadId
+      const capturedCwd = targetCwd || null
+      const capturedPrompt = nextText
+      void startTurnForThread(threadId, nextText, imageUrls, skills, fileAttachments, selectedMode)
+        .catch((unknownError) => {
+          shouldAutoScrollOnNextAgentEvent = false
+          setThreadInProgress(threadId, false)
+          setTurnActivityForThread(threadId, null)
+          const errorMessage = unknownError instanceof Error ? unknownError.message : 'Unknown application error'
+          setTurnErrorForThread(threadId, errorMessage)
+          error.value = errorMessage
+        })
+        .finally(() => {
+          isSendingMessage.value = false
+        })
+      void requestThreadTitleGeneration(capturedThreadId, capturedPrompt, capturedCwd)
+      return threadId
+    } catch (unknownError) {
+      shouldAutoScrollOnNextAgentEvent = false
+      if (threadId) {
+        setThreadInProgress(threadId, false)
+        setTurnActivityForThread(threadId, null)
+      }
+      const errorMessage = unknownError instanceof Error ? unknownError.message : 'Unknown application error'
+      if (threadId) {
+        setTurnErrorForThread(threadId, errorMessage)
+      }
+      error.value = errorMessage
+      isSendingMessage.value = false
+      throw unknownError
+    }
+  }
+
+  async function startTurnForThread(
+    threadId: string,
+    nextText: string,
+    imageUrls: string[] = [],
+    skills: Array<{ name: string; path: string }> = [],
+    fileAttachments: FileAttachment[] = [],
+    collaborationModeOverride?: CollaborationModeKind,
+  ): Promise<void> {
+    const reasoningEffort = selectedReasoningEffort.value
+    const collaborationMode = collaborationModeOverride === 'plan' ? 'plan' : collaborationModeOverride === 'default'
+      ? 'default'
+      : selectedCollaborationMode.value
+    const normalizedText = nextText.trim()
+    const normalizedImageUrls = [...imageUrls]
+    if (
+      normalizedImageUrls.length === 0
+      && shouldReuseAttachedImageFromPrompt(normalizedText)
+    ) {
+      const latestAttachedImageUrl = findLatestUserLocalImageUrl(threadId)
+      if (latestAttachedImageUrl) {
+        normalizedImageUrls.push(latestAttachedImageUrl)
+      }
+    }
+    const normalizedSkills = skills.map((skill) => ({ name: skill.name, path: skill.path }))
+    const normalizedFileAttachments = fileAttachments.map((file) => ({ ...file }))
+
+    setPendingTurnRequest(threadId, {
+      text: normalizedText,
+      imageUrls: [...normalizedImageUrls],
+      skills: normalizedSkills,
+      fileAttachments: normalizedFileAttachments,
+      effort: reasoningEffort,
+      collaborationMode,
+      fallbackRetried: false,
+    })
+
+    try {
+      if (resumedThreadById.value[threadId] !== true) {
+        const resumedThread = await resumeThread(threadId)
+        setThreadModelId(threadId, resumedThread.model)
+      }
+      const modelId = readModelIdForThread(threadId)
+
+      let startedTurnId = ''
+      try {
+        startedTurnId = await startThreadTurn(
+          threadId,
+          nextText,
+          normalizedImageUrls,
+          modelId || undefined,
+          reasoningEffort || undefined,
+          skills.length > 0 ? skills : undefined,
+          fileAttachments,
+          collaborationMode,
+        )
+      } catch (unknownError) {
+        if (modelId && modelId !== MODEL_FALLBACK_ID && isUnsupportedChatGptModelError(unknownError)) {
+          await applyFallbackModelSelection(threadId)
+          setPendingTurnRequest(threadId, {
+            text: normalizedText,
+            imageUrls: [...normalizedImageUrls],
+            skills: normalizedSkills,
+            fileAttachments: normalizedFileAttachments,
+            effort: reasoningEffort,
+            collaborationMode,
+            fallbackRetried: true,
+          })
+          startedTurnId = await startThreadTurn(
+            threadId,
+            nextText,
+            normalizedImageUrls,
+            MODEL_FALLBACK_ID,
+            reasoningEffort || undefined,
+            skills.length > 0 ? skills : undefined,
+            fileAttachments,
+            collaborationMode,
+          )
+        } else {
+          throw unknownError
+        }
+      }
+
+      if (startedTurnId) {
+        activeTurnIdByThreadId.value = {
+          ...activeTurnIdByThreadId.value,
+          [threadId]: startedTurnId,
+        }
+        maybeUnblockInterruptForActiveTurn(threadId, startedTurnId)
+      }
+
+      resumedThreadById.value = {
+        ...resumedThreadById.value,
+        [threadId]: true,
+      }
+
+      pendingThreadMessageRefresh.add(threadId)
+      await syncFromNotifications()
+      scheduleDelayedTurnSync(threadId)
+    } catch (unknownError) {
+      throw unknownError
+    }
+  }
+
+  async function processQueuedMessages(threadId: string): Promise<void> {
+    if (queueProcessingByThreadId.value[threadId] === true) return
+    queueProcessingByThreadId.value = {
+      ...queueProcessingByThreadId.value,
+      [threadId]: true,
+    }
+    try {
+      queuedMessagesByThreadId.value = await getThreadQueueState()
+    } catch {
+      // Backend queue state is optional during transient bridge failures.
+    } finally {
+      queueProcessingByThreadId.value = omitKey(queueProcessingByThreadId.value, threadId)
+    }
+  }
+
+  function scheduleQueueStateRefresh(threadId: string): void {
+    void processQueuedMessages(threadId)
+    if (typeof window === 'undefined') return
+    window.setTimeout(() => {
+      void processQueuedMessages(threadId)
+    }, 650)
+  }
+
+  async function interruptSelectedThreadTurn(): Promise<void> {
+    const threadId = selectedThreadId.value
+    if (!threadId) return
+    if (inProgressById.value[threadId] !== true) return
+    if (interruptBlockedUntilPersistedByThreadId.value[threadId] === true) return
+    let turnId = activeTurnIdByThreadId.value[threadId]
+    if (!turnId) {
+      const { activeTurnId } = await getThreadDetail(threadId)
+      turnId = activeTurnId
+      if (turnId) {
+        activeTurnIdByThreadId.value = {
+          ...activeTurnIdByThreadId.value,
+          [threadId]: turnId,
+        }
+      }
+    }
+    if (!turnId) {
+      throw new Error('Could not determine active turn id for interrupt')
+    }
+
+    isInterruptingTurn.value = true
+    error.value = ''
+    try {
+      await interruptThreadTurn(threadId, turnId)
+      setThreadInProgress(threadId, false)
+      setTurnActivityForThread(threadId, null)
+      setTurnErrorForThread(threadId, null)
+      if (activeTurnIdByThreadId.value[threadId]) {
+        activeTurnIdByThreadId.value = omitKey(activeTurnIdByThreadId.value, threadId)
+      }
+      pendingThreadMessageRefresh.add(threadId)
+      pendingThreadsRefresh = true
+      await syncFromNotifications()
+    } catch (unknownError) {
+      const errorMessage = unknownError instanceof Error ? unknownError.message : 'Failed to interrupt active turn'
+      setTurnErrorForThread(threadId, errorMessage)
+      error.value = errorMessage
+    } finally {
+      isInterruptingTurn.value = false
+    }
+  }
+
+  async function rollbackSelectedThread(turnId: string): Promise<void> {
+    const threadId = selectedThreadId.value
+    if (!threadId) return
+    if (isRollingBack.value) return
+    if (!turnId.trim()) return
+
+    const persisted = persistedMessagesByThreadId.value[threadId] ?? []
+    const matchedMessage = persisted.find((message) => message.turnId === turnId)
+    const turnIndex = typeof matchedMessage?.turnIndex === 'number' ? matchedMessage.turnIndex : -1
+    if (turnIndex < 0) return
+    const maxTurnIndex = persisted.reduce((max, m) => (typeof m.turnIndex === 'number' && m.turnIndex > max ? m.turnIndex : max), -1)
+    if (maxTurnIndex < 0 || turnIndex > maxTurnIndex) return
+    const numTurns = maxTurnIndex - turnIndex + 1
+    if (numTurns < 1) return
+
+    isRollingBack.value = true
+    error.value = ''
+    try {
+      const threadCwd = selectedThread.value?.cwd?.trim() ?? ''
+      if (threadCwd) {
+        await revertThreadFileChanges(threadId, turnId, threadCwd)
+      }
+      const nextMessages = await rollbackThread(threadId, numTurns)
+      setPersistedMessagesForThread(threadId, nextMessages)
+      setLiveAgentMessagesForThread(threadId, [])
+      clearLiveReasoningForThread(threadId)
+      if (liveCommandsByThreadId.value[threadId]) {
+        liveCommandsByThreadId.value = omitKey(liveCommandsByThreadId.value, threadId)
+      }
+      setTurnSummaryForThread(threadId, null)
+      setTurnActivityForThread(threadId, null)
+      setTurnErrorForThread(threadId, null)
+      pendingThreadsRefresh = true
+      await syncFromNotifications()
+    } catch (unknownError) {
+      error.value = unknownError instanceof Error ? unknownError.message : 'Failed to rollback thread'
+    } finally {
+      isRollingBack.value = false
+    }
+  }
+
+  let renameProjectTimer: ReturnType<typeof setTimeout> | null = null
+
+  async function persistProjectLabelToGlobalState(projectName: string, displayName: string): Promise<void> {
+    try {
+      const rootsState = await getWorkspaceRootsState()
+      const nextLabels = { ...rootsState.labels }
+      let changed = false
+      for (const rootPath of rootsState.order) {
+        if (!matchesWorkspaceRootProject(rootPath, projectName)) continue
+        const trimmed = displayName.trim()
+        if (trimmed.length === 0) {
+          if (nextLabels[rootPath] !== undefined) {
+            delete nextLabels[rootPath]
+            changed = true
+          }
+        } else if (nextLabels[rootPath] !== trimmed) {
+          nextLabels[rootPath] = trimmed
+          changed = true
+        }
+      }
+      if (changed) {
+        await setWorkspaceRootsState({
+          order: rootsState.order,
+          labels: nextLabels,
+          active: rootsState.active,
+          projectOrder: rootsState.projectOrder,
+        })
+      }
+    } catch {
+      // Keep localStorage-only rename when global state is unavailable.
+    }
+  }
+
+  function renameProject(projectName: string, displayName: string): void {
+    if (projectName.length === 0) return
+
+    const currentValue = projectDisplayNameById.value[projectName] ?? ''
+    if (currentValue === displayName) return
+
+    projectDisplayNameById.value = {
+      ...projectDisplayNameById.value,
+      [projectName]: displayName,
+    }
+    saveProjectDisplayNames(projectDisplayNameById.value)
+
+    if (renameProjectTimer !== null) clearTimeout(renameProjectTimer)
+    renameProjectTimer = setTimeout(() => {
+      renameProjectTimer = null
+      void persistProjectLabelToGlobalState(projectName, displayName)
+    }, 500)
+  }
+
+  async function removeProject(projectName: string): Promise<void> {
+    if (projectName.length === 0) return
+
+    const nextProjectOrder = projectOrder.value.filter((name) => name !== projectName)
+    if (!areStringArraysEqual(projectOrder.value, nextProjectOrder)) {
+      projectOrder.value = nextProjectOrder
+      saveProjectOrder(projectOrder.value)
+    }
+
+    sourceGroups.value = sourceGroups.value.filter((group) => group.projectName !== projectName)
+
+    if (projectDisplayNameById.value[projectName] !== undefined) {
+      const nextDisplayNames = { ...projectDisplayNameById.value }
+      delete nextDisplayNames[projectName]
+      projectDisplayNameById.value = nextDisplayNames
+      saveProjectDisplayNames(nextDisplayNames)
+    }
+
+    applyThreadFlags()
+
+    const flatThreads = flattenThreads(projectGroups.value)
+    pruneThreadScopedState(flatThreads)
+
+    const currentExists = flatThreads.some((thread) => thread.id === selectedThreadId.value)
+    if (!currentExists) {
+      setSelectedThreadId(flatThreads[0]?.id ?? '')
+    }
+
+    const removedRootPaths = new Set<string>()
+    try {
+      const rootsState = await getWorkspaceRootsState()
+      collectWorkspaceRootPathsForProjectRemoval(rootsState, projectName).forEach((rootPath) => {
+        removedRootPaths.add(rootPath)
+      })
+    } catch {
+      // Keep local-only removal when global state is unavailable.
+    }
+
+    if (removedRootPaths.size > 0) {
+      try {
+        const rootsState = await getWorkspaceRootsState()
+        const nextOrder = rootsState.order.filter((rootPath) => !removedRootPaths.has(rootPath))
+        const nextActive = rootsState.active.filter((rootPath) => !removedRootPaths.has(rootPath))
+        const fallbackActive = nextActive.length === 0 && nextOrder.length > 0
+          ? [nextOrder[0]]
+          : nextActive
+        await setWorkspaceRootsState({
+          order: nextOrder,
+          labels: omitKeys(rootsState.labels, removedRootPaths),
+          active: fallbackActive,
+          projectOrder: rootsState.projectOrder.filter((item) => item !== projectName && !removedRootPaths.has(item)),
+        })
+        return
+      } catch {
+        // Fall back to order-only persistence if direct removal fails.
+      }
+    }
+
+    await persistProjectOrderToWorkspaceRoots()
+  }
+
+  function reorderProject(projectName: string, toIndex: number): void {
+    if (projectName.length === 0) return
+    if (sourceGroups.value.length === 0) return
+
+    const visibleOrder = sourceGroups.value.map((group) => group.projectName)
+    const fromIndex = visibleOrder.indexOf(projectName)
+    if (fromIndex === -1) return
+
+    const clampedToIndex = Math.max(0, Math.min(toIndex, visibleOrder.length - 1))
+    const reorderedVisibleOrder = reorderStringArray(visibleOrder, fromIndex, clampedToIndex)
+    if (reorderedVisibleOrder === visibleOrder) return
+
+    const normalizedProjectOrder = mergeProjectOrder(reorderedVisibleOrder, sourceGroups.value)
+    projectOrder.value = normalizedProjectOrder
+    saveProjectOrder(projectOrder.value)
+
+    const orderedGroups = orderGroupsByProjectOrder(sourceGroups.value, projectOrder.value)
+    sourceGroups.value = mergeThreadGroups(sourceGroups.value, orderedGroups)
+    applyThreadFlags()
+    void persistProjectOrderToWorkspaceRoots()
+  }
+
+  function pinProjectToTop(projectName: string): void {
+    const normalizedName = projectName.trim()
+    if (!normalizedName) return
+    const nextOrder = [normalizedName, ...projectOrder.value.filter((name) => name !== normalizedName)]
+    if (areStringArraysEqual(projectOrder.value, nextOrder)) return
+    projectOrder.value = nextOrder
+    saveProjectOrder(projectOrder.value)
+
+    const orderedGroups = orderGroupsByProjectOrder(sourceGroups.value, projectOrder.value)
+    sourceGroups.value = mergeThreadGroups(sourceGroups.value, orderedGroups)
+    applyThreadFlags()
+    void persistProjectOrderToWorkspaceRoots()
+  }
+
+  async function persistProjectOrderToWorkspaceRoots(): Promise<void> {
+    try {
+      const rootsState = await getWorkspaceRootsState()
+      const nextState = buildWorkspaceRootsProjectOrderState(rootsState, projectOrder.value, sourceGroups.value)
+
+      await setWorkspaceRootsState({
+        order: nextState.order,
+        labels: rootsState.labels,
+        active: nextState.active,
+        projectOrder: nextState.projectOrder,
+      })
+    } catch {
+      // Keep local project order when global state persistence is unavailable.
+    }
+  }
+
+  async function syncThreadStatus(): Promise<void> {
+    if (isPolling.value) return
+    isPolling.value = true
+
+    try {
+      await loadThreads()
+
+      if (!selectedThreadId.value) return
+
+      const threadId = selectedThreadId.value
+      const currentVersion = currentThreadVersion(threadId)
+      const loadedVersion = loadedVersionByThreadId.value[threadId] ?? ''
+      const hasVersionChange = currentVersion.length > 0 && currentVersion !== loadedVersion
+      const isInProgress = inProgressById.value[threadId] === true
+
+      if (isInProgress || hasVersionChange) {
+        await loadMessages(threadId, { silent: true })
+      }
+    } catch {
+      // ignore poll failures and keep last known state
+    } finally {
+      isPolling.value = false
+    }
+  }
+
+  async function syncFromNotifications(): Promise<void> {
+    if (isPolling.value) {
+      if (typeof window !== 'undefined' && eventSyncTimer === null) {
+        eventSyncTimer = window.setTimeout(() => {
+          eventSyncTimer = null
+          void syncFromNotifications()
+        }, EVENT_SYNC_DEBOUNCE_MS)
+      }
+      return
+    }
+
+    isPolling.value = true
+
+    const shouldRefreshThreads = pendingThreadsRefresh
+    const threadIdsToRefresh = new Set(pendingThreadMessageRefresh)
+    pendingThreadsRefresh = false
+    pendingThreadMessageRefresh.clear()
+
+    try {
+      if (shouldRefreshThreads) {
+        await loadThreads()
+      }
+
+      const activeThreadId = selectedThreadId.value
+      if (!activeThreadId) return
+
+      const isActiveDirty = threadIdsToRefresh.has(activeThreadId)
+      const isInProgress = inProgressById.value[activeThreadId] === true
+      const currentVersion = currentThreadVersion(activeThreadId)
+      const loadedVersion = loadedVersionByThreadId.value[activeThreadId] ?? ''
+      const hasVersionChange = currentVersion.length > 0 && currentVersion !== loadedVersion
+
+      const shouldRefreshActiveThread =
+        hasVersionChange ||
+        (isInProgress && loadedMessagesByThreadId.value[activeThreadId] !== true) ||
+        (isActiveDirty && loadedMessagesByThreadId.value[activeThreadId] !== true) ||
+        (shouldRefreshThreads && loadedMessagesByThreadId.value[activeThreadId] !== true)
+
+      if (shouldRefreshActiveThread) {
+        await loadMessages(activeThreadId, { silent: true })
+      }
+    } catch {
+      // Keep UI stable on transient event sync failures.
+    } finally {
+      isPolling.value = false
+
+      if (
+        (pendingThreadsRefresh || pendingThreadMessageRefresh.size > 0) &&
+        typeof window !== 'undefined' &&
+        eventSyncTimer === null
+      ) {
+        eventSyncTimer = window.setTimeout(() => {
+          eventSyncTimer = null
+          void syncFromNotifications()
+        }, EVENT_SYNC_DEBOUNCE_MS)
+      }
+    }
+  }
+
+  async function recoverBridgeState(): Promise<void> {
+    await loadPendingServerRequestsFromBridge()
+    pendingThreadsRefresh = !hasLoadedThreads.value
+    if (
+      selectedThreadId.value &&
+      loadedMessagesByThreadId.value[selectedThreadId.value] !== true
+    ) {
+      pendingThreadMessageRefresh.add(selectedThreadId.value)
+    }
+    await syncFromNotifications()
+  }
+
+  function startPolling(): void {
+    if (typeof window === 'undefined') return
+
+    if (stopNotificationStream) return
+    void loadPendingServerRequestsFromBridge()
+    stopNotificationStream = subscribeCodexNotifications((notification) => {
+      if (notification.method === 'ready') {
+        clearAllTransientTurnErrors()
+        void recoverBridgeState()
+        return
+      }
+      applyRealtimeUpdates(notification)
+      queueEventDrivenSync(notification)
+    })
+  }
+
+  async function loadPendingServerRequestsFromBridge(): Promise<void> {
+    try {
+      const rows = await getPendingServerRequests()
+      const normalizedRequests = rows
+        .map((row) => normalizeServerRequest(row))
+        .filter((request): request is UiServerRequest => request !== null)
+      replacePendingServerRequests(normalizedRequests)
+    } catch {
+      // Keep UI usable when pending request endpoint is temporarily unavailable.
+    }
+  }
+
+  async function respondToPendingServerRequest(reply: UiServerRequestReply): Promise<boolean> {
+    try {
+      await replyToServerRequest(reply.id, {
+        result: reply.result,
+        error: reply.error,
+      })
+      removePendingServerRequestById(reply.id)
+      return true
+    } catch (unknownError) {
+      error.value = unknownError instanceof Error ? unknownError.message : 'Failed to reply to server request'
+      return false
+    }
+  }
+
+  function stopPolling(): void {
+    if (stopNotificationStream) {
+      stopNotificationStream()
+      stopNotificationStream = null
+    }
+
+    pendingThreadsRefresh = false
+    pendingThreadMessageRefresh.clear()
+    pendingTurnStartsById.clear()
+    if (eventSyncTimer !== null && typeof window !== 'undefined') {
+      window.clearTimeout(eventSyncTimer)
+      eventSyncTimer = null
+    }
+    if (rateLimitRefreshTimer !== null && typeof window !== 'undefined') {
+      window.clearTimeout(rateLimitRefreshTimer)
+      rateLimitRefreshTimer = null
+    }
+    if (threadListBackgroundTimer !== null && typeof window !== 'undefined') {
+      window.clearTimeout(threadListBackgroundTimer)
+      threadListBackgroundTimer = null
+    }
+    if (typeof window !== 'undefined') {
+      for (const timerId of delayedTurnSyncTimerByThreadId.values()) {
+        window.clearTimeout(timerId)
+      }
+    }
+    delayedTurnSyncTimerByThreadId.clear()
+    activeReasoningItemId = ''
+    shouldAutoScrollOnNextAgentEvent = false
+    persistedMessagesByThreadId.value = {}
+    livePlanMessagesByThreadId.value = {}
+    liveAgentMessagesByThreadId.value = {}
+    liveReasoningTextByThreadId.value = {}
+    liveCommandsByThreadId.value = {}
+    liveFileChangeMessagesByThreadId.value = {}
+    turnIndexByTurnIdByThreadId.value = {}
+    turnActivityByThreadId.value = {}
+    turnSummaryByThreadId.value = {}
+    turnErrorByThreadId.value = {}
+    activeTurnIdByThreadId.value = {}
+    interruptBlockedUntilPersistedByThreadId.value = {}
+    threadListedByServerById.value = {}
+    persistedUserMessageByThreadId.value = {}
+    queuedMessagesByThreadId.value = {}
+    queueProcessingByThreadId.value = {}
+    persistQueueState()
+    codexRateLimit.value = null
+    threadTokenUsageByThreadId.value = {}
+  }
+
+  const selectedThreadQueuedMessages = computed<QueuedMessage[]>(() => {
+    const threadId = selectedThreadId.value
+    if (!threadId) return []
+    return queuedMessagesByThreadId.value[threadId] ?? []
+  })
+
+  function removeQueuedMessage(messageId: string): void {
+    const threadId = selectedThreadId.value
+    if (!threadId) return
+    const queue = queuedMessagesByThreadId.value[threadId]
+    if (!queue) return
+    const next = queue.filter((m) => m.id !== messageId)
+    queuedMessagesByThreadId.value = next.length > 0
+      ? { ...queuedMessagesByThreadId.value, [threadId]: next }
+      : omitKey(queuedMessagesByThreadId.value, threadId)
+    persistQueueState()
+  }
+
+  function reorderQueuedMessage(draggedId: string, targetId: string): void {
+    const threadId = selectedThreadId.value
+    if (!threadId) return
+    const queue = queuedMessagesByThreadId.value[threadId]
+    if (!queue) return
+
+    const fromIndex = queue.findIndex((m) => m.id === draggedId)
+    const toIndex = queue.findIndex((m) => m.id === targetId)
+    if (fromIndex < 0 || toIndex < 0 || fromIndex === toIndex) return
+
+    const next = [...queue]
+    const [moved] = next.splice(fromIndex, 1)
+    next.splice(toIndex, 0, moved)
+    queuedMessagesByThreadId.value = {
+      ...queuedMessagesByThreadId.value,
+      [threadId]: next,
+    }
+    persistQueueState()
+  }
+
+  function steerQueuedMessage(messageId: string): void {
+    const threadId = selectedThreadId.value
+    if (!threadId) return
+    const queue = queuedMessagesByThreadId.value[threadId]
+    if (!queue) return
+    const msg = queue.find((m) => m.id === messageId)
+    if (!msg) return
+    removeQueuedMessage(messageId)
+    setSelectedCollaborationMode(msg.collaborationMode)
+    void sendMessageToSelectedThread(msg.text, msg.imageUrls, msg.skills, 'steer', msg.fileAttachments)
+  }
+
+  function primeSelectedThread(threadId: string): void {
+    setSelectedThreadId(threadId)
+  }
+
+  return {
+    projectGroups,
+    projectDisplayNameById,
+    selectedThread,
+    selectedThreadTokenUsage,
+    selectedThreadTerminalOpen,
+    isSelectedThreadInterruptPending,
+    selectedThreadServerRequests,
+    selectedLiveOverlay,
+    codexQuota,
+    selectedThreadId,
+    availableCollaborationModes,
+    availableModelIds,
+    selectedCollaborationMode,
+    selectedModelId,
+    selectedReasoningEffort,
+    selectedSpeedMode,
+    installedSkills,
+    accountRateLimitSnapshots,
+    messages,
+    hasMoreOlderMessages,
+    isLoadingThreads,
+    isThreadListFullyLoaded,
+    isLoadingMessages,
+    isLoadingOlderMessages,
+    isSendingMessage,
+    isInterruptingTurn,
+    isUpdatingSpeedMode,
+    isRollingBack,
+
+    error,
+    refreshAll,
+    refreshSkills,
+    selectThread,
+    loadMessages,
+    loadOlderMessages,
+    ensureThreadMessagesLoaded,
+    setThreadTerminalOpen,
+    toggleSelectedThreadTerminal,
+    archiveThreadById,
+    renameThreadById,
+    forkThreadById,
+    forkThreadFromTurn,
+    rollbackSelectedThread,
+
+    sendMessageToSelectedThread,
+    sendMessageToNewThread,
+    interruptSelectedThreadTurn,
+    selectedThreadQueuedMessages,
+    removeQueuedMessage,
+    reorderQueuedMessage,
+    steerQueuedMessage,
+    setSelectedCollaborationMode,
+    readModelIdForThread,
+    setSelectedModelIdForThread,
+    setSelectedModelId,
+
+    setSelectedReasoningEffort,
+    updateSelectedSpeedMode,
+    respondToPendingServerRequest,
+    renameProject,
+    removeProject,
+    reorderProject,
+    pinProjectToTop,
+    startPolling,
+    stopPolling,
+    primeSelectedThread,
+  }
+}
